refactor: split recorder/mod.rs into focused modules
Extract the 3500-line god module into four focused files:

- flush_loop.rs: DrainState, FlushStats, flush_once, run_flush_loop
- handle.rs: TelemetryHandle, RuntimeTelemetryHandle, spawn, InstrumentedSpawnGuard
- guard.rs: TelemetryGuard, TraceRuntimeCoreBuilder, WorkerHandle
- builder.rs: TracedRuntimeBuilder (typestate), TelemetryCore, TracedRuntime,
  assemble_processors, PipelineConfig, TelemetryRuntimeError

mod.rs retains only the glue: module declarations, re-exports,
register_hooks, attach_runtime, ControlCommand, and tests.

Each extracted module has a narrow interface and concentrated locality:
a bug in the flush loop is in flush_loop.rs, not somewhere in a
3500-line file. No behavioral changes.
