+++ dial9-tokio-telemetry/src/telemetry/recorder/builder.rs
@@ -1,0 +1,986 @@
+use crate::primitives::sync::Arc;
+use crate::primitives::sync::atomic::Ordering;
+#[cfg(feature = "cpu-profiling")]
+use crate::rate_limit::rate_limited;
+use crate::telemetry::writer::{RotatingWriter, TraceWriter};
+use std::path::PathBuf;
+use std::time::Duration;
+
+use super::event_writer::EventWriter;
+use super::flush_loop::run_flush_loop;
+use super::guard::{TelemetryGuard, WorkerHandle};
+use super::handle::TelemetryHandle;
+use super::shared_state::SharedState;
+use super::{ControlCommand, attach_runtime};
+
+/// Marker: no trace path has been set yet.
+#[derive(Debug)]
+#[non_exhaustive]
+pub struct NoTracePath;
+/// Marker: a trace path has been set.
+#[derive(Debug)]
+#[non_exhaustive]
+pub struct HasTracePath;
+
+/// Marker: no pipeline strategy has been chosen yet. From this state the
+/// builder can transition to either S3 (via `with_s3_uploader`) or a custom
+/// pipeline (via `with_custom_pipeline`).
+#[derive(Debug)]
+#[non_exhaustive]
+pub struct PipelineUnset;
+
+/// Marker: the S3 preset has been selected. `with_s3_client` is available
+/// to bind a pre-built client; `with_custom_pipeline` is not in scope.
+#[derive(Debug)]
+#[non_exhaustive]
+pub struct PipelineS3;
+
+/// Marker: a custom pipeline has been configured. No further pipeline
+/// methods are available.
+#[derive(Debug)]
+#[non_exhaustive]
+pub struct PipelineCustom;
+
+pub(super) enum PipelineConfig {
+    Unset,
+    #[cfg(feature = "worker-s3")]
+    S3(crate::background_task::S3PipelineUploader),
+    Custom(Vec<Box<dyn crate::background_task::SegmentProcessor>>),
+}
+
+/// Builder for configuring a traced Tokio runtime.
+pub struct TracedRuntimeBuilder<P = NoTracePath, M = PipelineUnset> {
+    pub(super) enabled: bool,
+    pub(super) task_tracking_enabled: bool,
+    pub(super) task_dump_config: Option<crate::telemetry::task_dump_config::TaskDumpConfig>,
+    pub(super) trace_path: Option<PathBuf>,
+    pub(super) runtime_name: Option<String>,
+    #[cfg(feature = "cpu-profiling")]
+    pub(super) cpu_profiling_config: Option<crate::telemetry::cpu_profile::CpuProfilingConfig>,
+    #[cfg(feature = "cpu-profiling")]
+    pub(super) sched_event_config: Option<crate::telemetry::cpu_profile::SchedEventConfig>,
+    pub(super) pipeline: PipelineConfig,
+    /// Static segment metadata to inject into every rotated segment's
+    /// header. The S3 preset populates this from `S3Config::as_metadata`
+    /// so traces stay self-describing.
+    pub(super) segment_metadata: Vec<(String, String)>,
+    pub(super) worker_poll_interval: Option<Duration>,
+    pub(super) worker_metrics_sink: Option<metrique_writer::BoxEntrySink>,
+    pub(super) _marker: std::marker::PhantomData<(P, M)>,
+}
+
+impl<P, M> std::fmt::Debug for TracedRuntimeBuilder<P, M> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("TracedRuntimeBuilder")
+            .finish_non_exhaustive()
+    }
+}
+
+// Methods available regardless of trace-path or pipeline state.
+impl<P, M> TracedRuntimeBuilder<P, M> {
+    /// Set to `false` to build a plain runtime with no telemetry
+    /// installed and a dummy [`TelemetryGuard`]. Defaults to `true`.
+    ///
+    /// Unlike [`TelemetryGuard::enable`]/[`TelemetryGuard::disable`]
+    /// (which toggle recording at runtime), this controls whether
+    /// telemetry hooks and threads are installed at all.
+    pub fn install(mut self, enabled: bool) -> Self {
+        self.enabled = enabled;
+        self
+    }
+
+    /// Enable or disable task spawn/terminate tracking.
+    pub fn with_task_tracking(mut self, enabled: bool) -> Self {
+        self.task_tracking_enabled = enabled;
+        self
+    }
+
+    /// Capture async backtraces at yield points for tasks that stay idle
+    /// longer than the configured threshold.
+    ///
+    /// Requires the `taskdump` crate feature to actually record events
+    pub fn with_task_dumps(
+        mut self,
+        config: crate::telemetry::task_dump_config::TaskDumpConfig,
+    ) -> Self {
+        if cfg!(not(feature = "taskdump")) {
+            tracing::warn!(
+                "taskdumps enabled but `taskdump` feature was not. No task dumps will be captured."
+            )
+        }
+        self.task_dump_config = Some(config);
+        self
+    }
+
+    /// Set a human-readable name for this runtime. Used in segment metadata
+    /// to map runtime indices to names for the trace viewer.
+    pub fn with_runtime_name(mut self, name: impl Into<String>) -> Self {
+        self.runtime_name = Some(name.into());
+        self
+    }
+
+    /// Set static metadata embedded as a `SegmentMetadata` event in every
+    /// sealed segment file. Read back during analysis and attached to every
+    /// Span.
+    ///
+    /// [`with_s3_uploader`](Self::with_s3_uploader) injects bucket /
+    /// service_name / instance_path / boot_id automatically; call this
+    /// method when using [`with_custom_pipeline`](Self::with_custom_pipeline)
+    /// (or no pipeline) and you still want those entries — or when you want
+    /// to override the preset's defaults.
+    ///
+    /// Repeated calls **replace** the metadata, matching how
+    /// `with_s3_uploader` overwrites on a second call. The last call wins,
+    /// so `with_segment_metadata` placed *after* `with_s3_uploader`
+    /// overrides the preset's injection.
+    pub fn with_segment_metadata(mut self, entries: Vec<(String, String)>) -> Self {
+        self.segment_metadata = entries;
+        self
+    }
+
+    /// Enable CPU profiling with the given configuration (Linux only).
+    #[cfg(feature = "cpu-profiling")]
+    pub fn with_cpu_profiling(
+        mut self,
+        config: crate::telemetry::cpu_profile::CpuProfilingConfig,
+    ) -> Self {
+        self.cpu_profiling_config = Some(config);
+        self
+    }
+
+    /// Enable per-worker scheduler event capture (Linux only).
+    #[cfg(feature = "cpu-profiling")]
+    pub fn with_sched_events(
+        mut self,
+        config: crate::telemetry::cpu_profile::SchedEventConfig,
+    ) -> Self {
+        self.sched_event_config = Some(config);
+        self
+    }
+
+    /// Set how often the background worker polls for sealed segments.
+    pub fn with_worker_poll_interval(mut self, interval: Duration) -> Self {
+        self.worker_poll_interval = Some(interval);
+        self
+    }
+
+    /// Set a metrics sink for the background worker.
+    pub fn with_worker_metrics_sink(mut self, sink: metrique_writer::BoxEntrySink) -> Self {
+        self.worker_metrics_sink = Some(sink);
+        self
+    }
+
+    /// Attach a new runtime to an existing telemetry session.
+    ///
+    /// This reuses the `SharedState`, flush thread, writer, and CPU profiler
+    /// from the original `TelemetryGuard`. Only the tokio callbacks are
+    /// registered on the new builder. The new runtime's workers get a unique
+    /// runtime index so their `WorkerId`s don't collide with existing runtimes.
+    pub fn build_and_attach_to_telemetry(
+        self,
+        mut builder: tokio::runtime::Builder,
+        guard: &TelemetryGuard,
+    ) -> std::io::Result<tokio::runtime::Runtime> {
+        let (Some(shared), Some(control_tx)) = (guard.shared(), guard.control_tx()) else {
+            // Disabled guard: produce a plain tokio runtime with no
+            // telemetry hooks so attaching still works gracefully.
+            return builder.build();
+        };
+        attach_runtime(
+            shared,
+            builder,
+            self.runtime_name,
+            control_tx,
+            self.task_tracking_enabled,
+        )
+    }
+
+    pub(super) fn into_state<Q, N>(self) -> TracedRuntimeBuilder<Q, N> {
+        TracedRuntimeBuilder {
+            enabled: self.enabled,
+            task_tracking_enabled: self.task_tracking_enabled,
+            task_dump_config: self.task_dump_config,
+            trace_path: self.trace_path,
+            runtime_name: self.runtime_name,
+            #[cfg(feature = "cpu-profiling")]
+            cpu_profiling_config: self.cpu_profiling_config,
+            #[cfg(feature = "cpu-profiling")]
+            sched_event_config: self.sched_event_config,
+            pipeline: self.pipeline,
+            segment_metadata: self.segment_metadata,
+            worker_poll_interval: self.worker_poll_interval,
+            worker_metrics_sink: self.worker_metrics_sink,
+            _marker: std::marker::PhantomData,
+        }
+    }
+}
+
+// Pipeline-strategy entry points: only available before a strategy has
+// been chosen, so the user picks S3 OR a custom pipeline, not both.
+impl<P> TracedRuntimeBuilder<P, PipelineUnset> {
+    /// Configure the S3 upload preset for sealed trace segments.
+    ///
+    /// The resulting pipeline is `[Gzip, S3]` (with `[Symbolize, ...]`
+    /// prepended when CPU profiling is enabled). After this call, only
+    /// [`with_s3_client`](TracedRuntimeBuilder::with_s3_client) and a
+    /// repeated [`with_s3_uploader`](TracedRuntimeBuilder::with_s3_uploader)
+    /// override are available — `with_custom_pipeline` is no longer in scope.
+    #[cfg(feature = "worker-s3")]
+    pub fn with_s3_uploader(
+        mut self,
+        config: crate::background_task::s3::S3Config,
+    ) -> TracedRuntimeBuilder<P, PipelineS3> {
+        self.segment_metadata = config
+            .as_metadata()
+            .map(|(k, v)| (k.to_string(), v.to_string()))
+            .collect();
+        self.pipeline = PipelineConfig::S3(crate::background_task::S3PipelineUploader::new(
+            config, None,
+        ));
+        self.into_state()
+    }
+
+    /// Configure a fully custom processor pipeline. The closure receives a
+    /// [`PipelineBuilder`](crate::background_task::PipelineBuilder); chain
+    /// methods like `.gzip()`, `.write_back()`, `.s3(cfg)` for built-ins
+    /// and `.pipe(processor)` for user-supplied processors.
+    ///
+    /// Mutually exclusive with [`with_s3_uploader`](Self::with_s3_uploader).
+    ///
+    /// This is the "full control" path: the resulting pipeline is exactly
+    /// what the closure builds, with nothing prepended or appended. In
+    /// particular, unlike the S3 preset, this path does **not**:
+    /// - auto-populate writer-side segment metadata — call
+    ///   [`with_segment_metadata`](Self::with_segment_metadata) if you want
+    ///   identity entries (service, host, etc.) embedded in trace files.
+    /// - auto-prepend the `Symbolize` step when CPU profiling is enabled.
+    ///   Chain
+    ///   [`.symbolize()`](crate::background_task::PipelineBuilder::symbolize)
+    ///   first if you want symbolized stack frames.
+    pub fn with_custom_pipeline<F>(mut self, build: F) -> TracedRuntimeBuilder<P, PipelineCustom>
+    where
+        F: FnOnce(
+            crate::background_task::PipelineBuilder,
+        ) -> crate::background_task::PipelineBuilder,
+    {
+        let pipeline = build(crate::background_task::PipelineBuilder::new());
+        self.pipeline = PipelineConfig::Custom(pipeline.into_processors());
+        self.into_state()
+    }
+}
+
+// S3 mode — once the S3 preset is chosen, only S3-specific tweaks remain.
+#[cfg(feature = "worker-s3")]
+impl<P> TracedRuntimeBuilder<P, PipelineS3> {
+    /// Provide a pre-built S3 client (for custom credentials or endpoints).
+    /// Replaces any client previously bound to the configured S3 uploader.
+    pub fn with_s3_client(mut self, client: aws_sdk_s3::Client) -> Self {
+        if let PipelineConfig::S3(ref mut uploader) = self.pipeline {
+            uploader.set_client(client);
+        }
+        self
+    }
+
+    /// Replace the configured S3 uploader. A client previously bound via
+    /// [`with_s3_client`](Self::with_s3_client) is carried over to the new
+    /// uploader so that call order between the two is irrelevant.
+    pub fn with_s3_uploader(mut self, config: crate::background_task::s3::S3Config) -> Self {
+        self.segment_metadata = config
+            .as_metadata()
+            .map(|(k, v)| (k.to_string(), v.to_string()))
+            .collect();
+        let carried = match &mut self.pipeline {
+            PipelineConfig::S3(uploader) => uploader.take_client(),
+            _ => None,
+        };
+        self.pipeline = PipelineConfig::S3(crate::background_task::S3PipelineUploader::new(
+            config, carried,
+        ));
+        self
+    }
+}
+
+impl<M> TracedRuntimeBuilder<NoTracePath, M> {
+    /// Set the trace output path. This transitions the builder to
+    /// `HasTracePath`, enabling `build()` and `build_and_start()`.
+    pub fn with_trace_path(
+        mut self,
+        path: impl Into<PathBuf>,
+    ) -> TracedRuntimeBuilder<HasTracePath, M> {
+        self.trace_path = Some(path.into());
+        self.into_state()
+    }
+
+    /// Build with a custom writer (for tests or `NullWriter`).
+    /// No background worker is spawned.
+    pub fn build_with_writer(
+        self,
+        builder: tokio::runtime::Builder,
+        writer: impl TraceWriter + 'static,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        self.into_state::<HasTracePath, M>()
+            .build_inner(builder, Box::new(writer))
+    }
+
+    /// Build with a custom writer and immediately enable recording.
+    pub fn build_and_start_with_writer(
+        self,
+        builder: tokio::runtime::Builder,
+        writer: impl TraceWriter + 'static,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        let (runtime, guard) = self.build_with_writer(builder, writer)?;
+        guard.enable();
+        Ok((runtime, guard))
+    }
+
+    /// Build the traced runtime. No background worker is spawned
+    /// (use `with_trace_path()` first for worker support).
+    pub fn build(
+        self,
+        builder: tokio::runtime::Builder,
+        writer: impl TraceWriter + 'static,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        self.build_with_writer(builder, writer)
+    }
+
+    /// Build and immediately enable recording.
+    pub fn build_and_start(
+        self,
+        builder: tokio::runtime::Builder,
+        writer: impl TraceWriter + 'static,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        self.build_and_start_with_writer(builder, writer)
+    }
+}
+
+impl<M> TracedRuntimeBuilder<HasTracePath, M> {
+    /// Set the trace output path (no-op, already set).
+    pub fn with_trace_path(mut self, path: impl Into<PathBuf>) -> Self {
+        self.trace_path = Some(path.into());
+        self
+    }
+
+    /// Build the traced runtime with a `RotatingWriter`.
+    ///
+    /// The background worker is auto-spawned when cpu-profiling or any
+    /// pipeline strategy is configured. Recording starts disabled; call
+    /// [`TelemetryGuard::enable`] to begin, or use
+    /// [`build_and_start`](Self::build_and_start).
+    pub fn build(
+        self,
+        builder: tokio::runtime::Builder,
+        writer: RotatingWriter,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        self.build_inner(builder, Box::new(writer))
+    }
+
+    /// Build the traced runtime and immediately enable recording.
+    pub fn build_and_start(
+        self,
+        builder: tokio::runtime::Builder,
+        writer: RotatingWriter,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        let (runtime, guard) = self.build(builder, writer)?;
+        guard.enable();
+        Ok((runtime, guard))
+    }
+
+    /// Build with a custom writer (for tests). The background worker is
+    /// still spawned if cpu-profiling or any pipeline strategy is configured
+    /// and `trace_path` is set.
+    pub fn build_with_writer(
+        self,
+        builder: tokio::runtime::Builder,
+        writer: impl TraceWriter + 'static,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        self.build_inner(builder, Box::new(writer))
+    }
+
+    /// Build with a custom writer and immediately enable recording.
+    pub fn build_and_start_with_writer(
+        self,
+        builder: tokio::runtime::Builder,
+        writer: impl TraceWriter + 'static,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        let (runtime, guard) = self.build_with_writer(builder, writer)?;
+        guard.enable();
+        Ok((runtime, guard))
+    }
+
+    fn build_inner(
+        self,
+        builder: tokio::runtime::Builder,
+        writer: Box<dyn TraceWriter>,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        if !self.enabled {
+            return TracedRuntime::build_disabled(builder);
+        }
+
+        let processors = assemble_processors(
+            #[cfg(feature = "cpu-profiling")]
+            self.cpu_profiling_config.is_some(),
+            self.pipeline,
+        );
+
+        let core_builder = TelemetryCore::builder()
+            .writer(writer)
+            .maybe_trace_path(self.trace_path)
+            .maybe_task_dump_config(self.task_dump_config)
+            .maybe_worker_poll_interval(self.worker_poll_interval)
+            .maybe_worker_metrics_sink(self.worker_metrics_sink)
+            .processors(processors)
+            .segment_metadata(self.segment_metadata);
+
+        #[cfg(feature = "cpu-profiling")]
+        let core_builder = core_builder
+            .maybe_cpu_profiling(self.cpu_profiling_config)
+            .maybe_sched_events(self.sched_event_config);
+
+        let guard = core_builder.build()?;
+        let control_tx = guard
+            .control_tx()
+            .expect("TelemetryCore::builder().build() always returns an enabled guard")
+            .clone();
+        let shared = guard
+            .shared()
+            .expect("TelemetryCore::builder().build() always returns an enabled guard");
+        let runtime = attach_runtime(
+            shared,
+            builder,
+            self.runtime_name,
+            &control_tx,
+            self.task_tracking_enabled,
+        )?;
+        Ok((runtime, guard))
+    }
+}
+
+/// Build the final processor pipeline.
+///
+/// `Symbolize` is auto-prepended for the built-in presets (`Unset`, `S3`)
+/// when CPU profiling is enabled. The `Custom` path is "full control" — the
+/// user's processor list is passed through verbatim, and they're expected to
+/// chain [`PipelineBuilder::symbolize`](crate::background_task::PipelineBuilder::symbolize)
+/// themselves if they want symbolization.
+///
+/// Behaviour matrix:
+///
+/// | strategy | CPU profiling on               | CPU profiling off |
+/// |----------|--------------------------------|-------------------|
+/// | Unset    | `[Symbolize, Gzip, WriteBack]` | (worker skipped)  |
+/// | S3       | `[Symbolize, Gzip, S3]`        | `[Gzip, S3]`      |
+/// | Custom   | `[...user]`                    | `[...user]`       |
+pub(super) fn assemble_processors(
+    #[cfg(feature = "cpu-profiling")] cpu_profiling_enabled: bool,
+    pipeline: PipelineConfig,
+) -> Vec<Box<dyn crate::background_task::SegmentProcessor>> {
+    #[cfg(not(feature = "cpu-profiling"))]
+    let cpu_profiling_enabled = false;
+
+    if matches!(pipeline, PipelineConfig::Unset) && !cpu_profiling_enabled {
+        return Vec::new();
+    }
+
+    let mut processors: Vec<Box<dyn crate::background_task::SegmentProcessor>> = Vec::new();
+    match pipeline {
+        PipelineConfig::Unset => {
+            #[cfg(feature = "cpu-profiling")]
+            if cpu_profiling_enabled {
+                processors.push(Box::new(crate::background_task::SymbolizeProcessor));
+            }
+            processors.push(Box::new(crate::background_task::GzipCompressor));
+            processors.push(Box::new(crate::background_task::WriteBackProcessor));
+        }
+        #[cfg(feature = "worker-s3")]
+        PipelineConfig::S3(uploader) => {
+            #[cfg(feature = "cpu-profiling")]
+            if cpu_profiling_enabled {
+                processors.push(Box::new(crate::background_task::SymbolizeProcessor));
+            }
+            processors.push(Box::new(crate::background_task::GzipCompressor));
+            processors.push(Box::new(uploader));
+        }
+        PipelineConfig::Custom(user) => {
+            processors.extend(user);
+        }
+    }
+    processors
+}
+
+/// Entry point for creating a telemetry session decoupled from any tokio runtime.
+///
+/// Use [`TelemetryCore::builder()`] to configure the session, then call
+/// [`TelemetryGuard::trace_runtime`] to attach one or more runtimes.
+///
+/// ```rust,no_run
+/// # use dial9_tokio_telemetry::telemetry::{RotatingWriter, TelemetryCore};
+/// # fn main() -> std::io::Result<()> {
+/// let writer = RotatingWriter::single_file("/tmp/trace.bin")?;
+/// let guard = TelemetryCore::builder()
+///     .writer(writer)
+///     .build()?;
+/// guard.enable();
+///
+/// let mut builder = tokio::runtime::Builder::new_multi_thread();
+/// builder.worker_threads(4).enable_all();
+/// let (runtime, handle) = guard.trace_runtime("main").build(builder)?;
+/// # Ok(())
+/// # }
+/// ```
+#[derive(Debug)]
+pub struct TelemetryCore;
+
+#[bon::bon]
+impl TelemetryCore {
+    /// Build a telemetry session. Recording starts disabled; call
+    /// [`TelemetryGuard::enable`] to begin recording.
+    #[builder(state_mod = telemetry_core_builder)]
+    pub fn new(
+        /// The pipeline of [`SegmentProcessor`](crate::background_task::SegmentProcessor)s
+        /// to run on each sealed segment. When empty the background worker
+        /// is not spawned.
+        #[builder(field)]
+        processors: Vec<Box<dyn crate::background_task::SegmentProcessor>>,
+        /// Static segment metadata injected into every rotated segment's
+        /// header. Empty by default; the S3 preset populates it from the
+        /// configured `S3Config` so traces stay self-describing.
+        #[builder(field)]
+        segment_metadata: Vec<(String, String)>,
+        /// S3 upload configuration.
+        #[cfg(feature = "worker-s3")]
+        #[builder(field)]
+        s3_config: Option<crate::background_task::s3::S3Config>,
+        /// Pre-built S3 client.
+        #[cfg(feature = "worker-s3")]
+        #[builder(field)]
+        s3_client: Option<aws_sdk_s3::Client>,
+        /// The trace writer (e.g. [`RotatingWriter`], [`NullWriter`](crate::telemetry::NullWriter)).
+        writer: impl TraceWriter + 'static,
+        /// Path for trace output. Enables the background worker when any
+        /// segment processors are configured.
+        #[builder(into)]
+        trace_path: Option<PathBuf>,
+        /// Capture async backtraces at yield points. Requires the `taskdump`
+        /// crate feature to actually record events.
+        task_dump_config: Option<crate::telemetry::task_dump_config::TaskDumpConfig>,
+        /// Enable CPU profiling (Linux only).
+        #[cfg(feature = "cpu-profiling")]
+        cpu_profiling: Option<crate::telemetry::cpu_profile::CpuProfilingConfig>,
+        /// Enable scheduler event capture (Linux only).
+        #[cfg(feature = "cpu-profiling")]
+        sched_events: Option<crate::telemetry::cpu_profile::SchedEventConfig>,
+        /// How often the background worker polls for sealed segments.
+        worker_poll_interval: Option<Duration>,
+        /// Metrics sink for the flush/worker threads.
+        worker_metrics_sink: Option<metrique_writer::BoxEntrySink>,
+    ) -> std::io::Result<TelemetryGuard> {
+        let start_mono_ns = crate::telemetry::events::clock_monotonic_ns();
+        let rng_seed = task_dump_config.as_ref().and_then(|cfg| cfg.rng_seed());
+        let shared = Arc::new(SharedState::new(start_mono_ns, rng_seed));
+        if let Some(cfg) = task_dump_config.as_ref() {
+            shared.task_dumps_enabled.store(true, Ordering::Relaxed);
+            shared
+                .task_dump_idle_threshold_ns
+                .store(cfg.idle_threshold().as_nanos() as u64, Ordering::Relaxed);
+        }
+
+        // Determine the pipeline strategy from the builder fields, then
+        // delegate to `assemble_processors` — the single source of truth for
+        // which processors are used in each configuration.
+        #[allow(unused_mut)]
+        let mut segment_metadata = segment_metadata;
+
+        #[allow(unused_variables)]
+        let pipeline = if !processors.is_empty() {
+            PipelineConfig::Custom(processors)
+        } else {
+            #[cfg(feature = "worker-s3")]
+            if let Some(config) = s3_config {
+                if segment_metadata.is_empty() {
+                    segment_metadata = config
+                        .as_metadata()
+                        .map(|(k, v)| (k.to_string(), v.to_string()))
+                        .collect();
+                }
+                PipelineConfig::S3(crate::background_task::S3PipelineUploader::new(
+                    config, s3_client,
+                ))
+            } else {
+                PipelineConfig::Unset
+            }
+            #[cfg(not(feature = "worker-s3"))]
+            PipelineConfig::Unset
+        };
+
+        let processors = assemble_processors(
+            #[cfg(feature = "cpu-profiling")]
+            cpu_profiling.is_some(),
+            pipeline,
+        );
+
+        #[allow(unused_mut)]
+        let mut event_writer = EventWriter::new(Box::new(writer));
+
+        if !segment_metadata.is_empty() {
+            event_writer.update_segment_metadata(segment_metadata);
+        }
+
+        #[cfg(feature = "cpu-profiling")]
+        {
+            if let Some(ref config) = cpu_profiling {
+                match crate::telemetry::cpu_profile::CpuProfiler::start(config.clone()) {
+                    Ok(sampler) => shared.push_source(Box::new(sampler)),
+                    Err(e) => rate_limited!(Duration::from_secs(60), {
+                        tracing::warn!("failed to start CPU profiler: {e}");
+                    }),
+                }
+            }
+            if let Some(sched_cfg) = sched_events {
+                match crate::telemetry::cpu_profile::SchedProfiler::new(sched_cfg) {
+                    Ok(sched) => shared.push_source(Box::new(sched)),
+                    Err(e) => rate_limited!(Duration::from_secs(60), {
+                        tracing::warn!("failed to start scheduler event profiler: {e}");
+                    }),
+                }
+            }
+        }
+
+        // Channel for TelemetryHandle/Guard → flush thread communication.
+        let (control_tx, control_rx) =
+            crate::primitives::sync::mpsc::sync_channel::<ControlCommand>(1);
+
+        let flush_metrics_sink = worker_metrics_sink
+            .clone()
+            .unwrap_or_else(metrique_writer::sink::DevNullSink::boxed);
+
+        let flush_thread = {
+            let shared = shared.clone();
+            crate::primitives::thread::spawn_named("dial9-flush", move || {
+                #[cfg(target_os = "linux")]
+                // SAFETY: nice() is a simple syscall with no memory safety
+                // implications. Increasing the nice value (lowering priority)
+                // is always permitted for unprivileged processes.
+                unsafe {
+                    let _ = libc::nice(10);
+                }
+
+                #[cfg(feature = "cpu-profiling")]
+                let _ = dial9_perf_self_profile::register_current_thread();
+                run_flush_loop(control_rx, &shared, &flush_metrics_sink, event_writer);
+                #[cfg(feature = "cpu-profiling")]
+                dial9_perf_self_profile::unregister_current_thread();
+            })
+        };
+
+        // Auto-construct worker config when we have a trace path and
+        // at least one processor. When the user supplies no processors
+        // there is nothing for the worker to do, so skip spawning it.
+        let worker_config = trace_path.and_then(|trace_path| {
+            if processors.is_empty() {
+                return None;
+            }
+
+            let poll_interval =
+                worker_poll_interval.unwrap_or(crate::background_task::DEFAULT_POLL_INTERVAL);
+            let metrics_sink =
+                worker_metrics_sink.unwrap_or_else(metrique_writer::sink::DevNullSink::boxed);
+
+            Some(
+                crate::background_task::BackgroundTaskConfig::builder()
+                    .trace_path(trace_path)
+                    .poll_interval(poll_interval)
+                    .processors(processors)
+                    .metrics_sink(metrics_sink)
+                    .build(),
+            )
+        });
+
+        #[allow(unused_mut)]
+        let mut worker = None;
+        if let Some(config) = worker_config {
+            let (shutdown_tx, shutdown_rx) = tokio::sync::oneshot::channel();
+            let wt = crate::primitives::thread::spawn_named("dial9-worker", move || {
+                #[cfg(feature = "cpu-profiling")]
+                let _ = dial9_perf_self_profile::register_current_thread();
+                crate::background_task::run_background_task(config, shutdown_rx);
+                #[cfg(feature = "cpu-profiling")]
+                dial9_perf_self_profile::unregister_current_thread();
+            });
+            worker = Some(WorkerHandle {
+                shutdown: Some(shutdown_tx),
+                thread: Some(wt),
+            });
+        }
+
+        Ok(TelemetryGuard::enabled(
+            TelemetryHandle::enabled(shared, control_tx),
+            Some(flush_thread),
+            worker,
+        ))
+    }
+}
+
+// Custom methods on the generated builder.
+impl<W: TraceWriter, S: telemetry_core_builder::State> TelemetryCoreBuilder<W, S> {
+    /// Configure S3 upload for sealed trace segments.
+    #[cfg(feature = "worker-s3")]
+    pub fn s3_config(mut self, config: crate::background_task::s3::S3Config) -> Self {
+        self.s3_config = Some(config);
+        self
+    }
+
+    /// Provide a pre-built S3 client (for custom credentials or endpoints).
+    #[cfg(feature = "worker-s3")]
+    pub fn s3_client(mut self, client: aws_sdk_s3::Client) -> Self {
+        self.s3_client = Some(client);
+        self
+    }
+
+    /// Set the processor pipeline directly.
+    pub fn processors(
+        mut self,
+        processors: Vec<Box<dyn crate::background_task::SegmentProcessor>>,
+    ) -> Self {
+        self.processors = processors;
+        self
+    }
+
+    /// Set static segment metadata.
+    pub fn segment_metadata(mut self, entries: Vec<(String, String)>) -> Self {
+        self.segment_metadata = entries;
+        self
+    }
+}
+
+/// A tokio runtime paired with its (optional) dial9 telemetry guard.
+///
+/// The guard, when present, must outlive the runtime so traces are flushed
+/// on drop — keeping both inside one struct enforces that ordering at the
+/// type level (fields drop top-to-bottom, so `runtime` drops before `guard`).
+///
+/// Construct one of two ways:
+///
+/// - **High-level**: from a [`crate::Dial9Config`] via [`TracedRuntime::new`]
+///   (panicking, used by the `#[dial9_tokio_telemetry::main]` macro) or
+///   [`TracedRuntime::try_new`] (fallible).
+/// - **Low-level**: via [`TracedRuntime::builder`] →
+///   [`TracedRuntimeBuilder::build_and_start`] for direct control over the
+///   raw [`tokio::runtime::Builder`] and [`crate::telemetry::TraceWriter`].
+///   This is the path used by example code, benchmarks, and integration
+///   tests that want to wire a [`crate::telemetry::NullWriter`] or other
+///   custom writer.
+#[derive(Debug)]
+pub struct TracedRuntime {
+    pub(crate) runtime: tokio::runtime::Runtime,
+    pub(crate) guard: TelemetryGuard,
+}
+
+impl TracedRuntime {
+    /// Create a new [`TracedRuntimeBuilder`].
+    pub fn builder() -> TracedRuntimeBuilder<NoTracePath, PipelineUnset> {
+        TracedRuntimeBuilder {
+            enabled: true,
+            task_tracking_enabled: false,
+            task_dump_config: None,
+            trace_path: None,
+            runtime_name: None,
+            #[cfg(feature = "cpu-profiling")]
+            cpu_profiling_config: None,
+            #[cfg(feature = "cpu-profiling")]
+            sched_event_config: None,
+            pipeline: PipelineConfig::Unset,
+            segment_metadata: Vec::new(),
+            worker_poll_interval: None,
+            worker_metrics_sink: None,
+            _marker: std::marker::PhantomData,
+        }
+    }
+
+    /// Build a plain runtime with no telemetry installed.
+    ///
+    /// The returned [`TelemetryGuard`] is in its disabled mode — see
+    /// [`TelemetryGuard::is_enabled`].
+    pub fn build_disabled(
+        mut builder: tokio::runtime::Builder,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        let runtime = builder.build()?;
+        Ok((runtime, TelemetryGuard::disabled()))
+    }
+
+    /// Build the traced runtime. Recording starts disabled.
+    pub fn build(
+        builder: tokio::runtime::Builder,
+        writer: impl TraceWriter + 'static,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        Self::builder().build_with_writer(builder, writer)
+    }
+
+    /// Build the traced runtime and immediately enable recording.
+    pub fn build_and_start(
+        builder: tokio::runtime::Builder,
+        writer: impl TraceWriter + 'static,
+    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
+        Self::builder().build_and_start_with_writer(builder, writer)
+    }
+}
+
+// ---------------------------------------------------------------------------
+// High-level construction: TracedRuntime::new / try_new from Dial9Config
+// ---------------------------------------------------------------------------
+
+/// Errors produced while constructing a [`TracedRuntime`] from a
+/// [`crate::Dial9Config`].
+///
+/// Writer-transport I/O has already been validated by
+/// [`crate::Dial9ConfigBuilder::build`], so the only remaining failure
+/// modes here come from the tokio runtime builder and the telemetry
+/// background worker startup.
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum TelemetryRuntimeError {
+    /// Failure from [`tokio::runtime::Builder::build`].
+    TokioRuntimeBuilder(std::io::Error),
+    /// Failure from telemetry core setup (traced runtime + background worker).
+    TelemetryCore(std::io::Error),
+}
+
+impl std::fmt::Display for TelemetryRuntimeError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            TelemetryRuntimeError::TokioRuntimeBuilder(e) => {
+                write!(f, "tokio runtime builder: {e}")
+            }
+            TelemetryRuntimeError::TelemetryCore(e) => write!(f, "telemetry core: {e}"),
+        }
+    }
+}
+
+impl std::error::Error for TelemetryRuntimeError {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        match self {
+            TelemetryRuntimeError::TokioRuntimeBuilder(e)
+            | TelemetryRuntimeError::TelemetryCore(e) => Some(e),
+        }
+    }
+}
+
+/// Drive a [`crate::current_config::Inner`] to a tokio runtime + guard.
+///
+/// `Inner::Enabled` already carries a built [`RotatingWriter`], so this
+/// only needs to materialize the tokio builder and hand both off to
+/// [`TracedRuntimeBuilder::build_and_start`]. `Inner::Disabled`
+/// produces a plain tokio runtime paired with a disabled
+/// [`TelemetryGuard`].
+fn try_assemble_dial9_config(
+    inner: crate::current_config::Inner,
+) -> Result<(tokio::runtime::Runtime, TelemetryGuard), TelemetryRuntimeError> {
+    use crate::current_config::{Inner, materialize_tokio_builder};
+
+    match inner {
+        Inner::Enabled {
+            writer,
+            tokio_configurators,
+            runtime_builder,
+        } => {
+            let tokio_builder = materialize_tokio_builder(&tokio_configurators);
+            let (runtime, guard) = runtime_builder
+                .build_and_start(tokio_builder, writer)
+                .map_err(TelemetryRuntimeError::TelemetryCore)?;
+            Ok((runtime, guard))
+        }
+        Inner::Disabled {
+            tokio_configurators,
+        } => {
+            let runtime = materialize_tokio_builder(&tokio_configurators)
+                .build()
+                .map_err(TelemetryRuntimeError::TokioRuntimeBuilder)?;
+            Ok((runtime, TelemetryGuard::disabled()))
+        }
+    }
+}
+
+impl TracedRuntime {
+    /// Build a [`TracedRuntime`] from a config, panicking with the
+    /// underlying error on failure. Used by the
+    /// `#[dial9_tokio_telemetry::main]` macro.
+    ///
+    /// Generic over any input that converts into a [`TracedRuntime`]: in
+    /// practice that means either the fluent
+    /// [`crate::Dial9Config`] (returned by
+    /// [`Dial9Config::builder`](crate::Dial9Config::builder)) or the
+    /// deprecated positional [`crate::config::Dial9Config`].
+    ///
+    /// # Panics
+    ///
+    /// Panics if the underlying conversion fails.
+    ///
+    /// For fallible construction, use [`try_new`](Self::try_new).
+    pub fn new<C>(config: C) -> Self
+    where
+        C: TryInto<TracedRuntime>,
+        <C as TryInto<TracedRuntime>>::Error: std::fmt::Display,
+    {
+        config
+            .try_into()
+            .unwrap_or_else(|e| panic!("failed to initialize runtime: {e}"))
+    }
+
+    /// Fallible counterpart to [`new`](Self::new).
+    pub fn try_new<C>(config: C) -> Result<Self, <C as TryInto<TracedRuntime>>::Error>
+    where
+        C: TryInto<TracedRuntime>,
+    {
+        config.try_into()
+    }
+
+    /// Borrow the underlying tokio runtime.
+    pub fn runtime(&self) -> &tokio::runtime::Runtime {
+        &self.runtime
+    }
+
+    /// Borrow the telemetry guard.
+    pub fn guard(&self) -> &TelemetryGuard {
+        &self.guard
+    }
+
+    /// Run `fut` to completion on the runtime.
+    pub fn block_on<F>(&self, fut: F) -> F::Output
+    where
+        F: std::future::Future + Send + 'static,
+        F::Output: Send + 'static,
+    {
+        let handle = self.guard.handle();
+        self.runtime.block_on(async move {
+            match handle.spawn(fut).await {
+                Ok(output) => output,
+                Err(err) if err.is_panic() => std::panic::resume_unwind(err.into_panic()),
+                Err(_) => unreachable!("task cannot be cancelled inside block_on"),
+            }
+        })
+    }
+}
+
+impl TryFrom<crate::Dial9Config> for TracedRuntime {
+    type Error = TelemetryRuntimeError;
+
+    fn try_from(config: crate::Dial9Config) -> Result<Self, Self::Error> {
+        let (runtime, guard) = try_assemble_dial9_config(config.0)?;
+        Ok(Self { runtime, guard })
+    }
+}
+
+/// Bridge for the deprecated positional config API at
+/// [`crate::config::Dial9Config`] so that it remains compatible with
+/// [`TracedRuntime::new`] (and therefore the
+/// `#[dial9_tokio_telemetry::main]` macro).
+impl TryFrom<crate::config::Dial9Config> for TracedRuntime {
+    type Error = std::io::Error;
+
+    fn try_from(config: crate::config::Dial9Config) -> Result<Self, Self::Error> {
+        let (runtime, guard) = config.build()?;
+        Ok(Self {
+            runtime,
+            guard: guard.unwrap_or_else(TelemetryGuard::disabled),
+        })
+    }
+}
