+++ dial9-tokio-telemetry/src/telemetry/recorder/mod.rs
@@ -1,43 +1,43 @@
+mod builder;
 mod event_writer;
+mod flush_loop;
+mod guard;
+mod handle;
 mod runtime_context;
 mod shared_state;
 pub(crate) mod source;
 
 pub(crate) use runtime_context::RuntimeContext;
 pub use runtime_context::current_worker_id;
+pub(crate) use runtime_context::poll_start_ts_monotonic;
 pub(crate) use shared_state::SharedState;
 
-use event_writer::EventWriter;
+pub use builder::{
+    HasTracePath, NoTracePath, PipelineCustom, PipelineS3, PipelineUnset, TelemetryCore,
+    TelemetryCoreBuilder, TelemetryRuntimeError, TracedRuntime, TracedRuntimeBuilder,
+};
+pub use guard::{TelemetryGuard, TraceRuntimeCoreBuilder};
+pub use handle::{RuntimeTelemetryHandle, TelemetryHandle, spawn};
+
+pub(crate) use flush_loop::FlushStats;
+
+// Re-exports for internal test access
+#[cfg(test)]
+pub(self) use builder::PipelineConfig;
+#[cfg(test)]
+pub(self) use event_writer::EventWriter;
+#[cfg(test)]
+pub(self) use handle::InstrumentedSpawnGuard;
+
+use handle::{CURRENT_HANDLE, INSTRUMENTED_SPAWN};
 use runtime_context::{make_poll_end, make_poll_start, make_worker_park, make_worker_unpark};
 
-use crate::metrics::{FlushMetrics, Operation, TlDrainMetrics};
 use crate::primitives::sync::Arc;
 use crate::primitives::sync::atomic::Ordering;
 use crate::rate_limit::rate_limited;
-use crate::telemetry::buffer;
 use crate::telemetry::format::TaskTerminateEvent;
 use crate::telemetry::task_metadata::TaskId;
-use crate::telemetry::writer::{RotatingWriter, TraceWriter};
-use metrique::timers::Timer;
-use metrique::unit::Microsecond;
-use metrique::unit_of_work::metrics;
-use metrique_timesource::time_source;
-use std::cell::Cell;
-use std::cell::RefCell;
-use std::path::PathBuf;
 use std::time::Duration;
-
-pub(crate) use runtime_context::poll_start_ts_monotonic;
-
-crate::primitives::thread_local! {
-    /// Per-thread [`TelemetryHandle`], populated in `on_thread_start` and
-    /// cleared in `on_thread_stop`. Enables [`TelemetryHandle::current`].
-    static CURRENT_HANDLE: RefCell<Option<TelemetryHandle>> = const { RefCell::new(None) };
-
-    /// Nest count for [`InstrumentedSpawnGuard`]. `on_task_spawn` treats
-    /// any value `> 0` as an instrumented spawn.
-    static INSTRUMENTED_SPAWN: Cell<u32> = const { Cell::new(0) };
-}
 
 // ---------------------------------------------------------------------------
 // Channel-based control for the flush thread
@@ -47,96 +47,6 @@ crate::primitives::thread_local! {
 pub(crate) enum ControlCommand {
     /// Flush, finalize (seal segment), then exit the thread.
     FinalizeAndStop(crate::primitives::sync::mpsc::SyncSender<()>),
-}
-
-/// Tracks the drain coordination state between the flush loop and the writer.
-///
-/// When the writer reports a drain is due (`should_drain()`), we can't act
-/// immediately because thread-local buffers may still hold events that belong
-/// in the current segment. Instead we bump the drain epoch (so threads
-/// self-flush on their next `record_event`), wait one cycle (~5 ms) for that
-/// to propagate, then perform the intrusive drain + flush + notify the writer
-/// via `drained()`.
-///
-/// Without a state machine, the naïve check `if should_drain { schedule drain }`
-/// fires every cycle (since we haven't drained yet), forever deferring the
-/// actual drain.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum DrainState {
-    /// Normal operation — poll `should_drain()` each cycle.
-    Idle,
-    /// The writer reported drain due and we bumped the drain epoch.
-    /// Next cycle: intrusive drain + flush + `drained()`.
-    EpochBumped,
-}
-
-/// Stats returned by flush for metrics publishing.
-#[metrics(subfield, rename_all = "PascalCase")]
-#[derive(Debug)]
-pub(crate) struct FlushStats {
-    pub event_count: u64,
-    pub dropped_batches: u64,
-    #[metrics(unit = Microsecond)]
-    pub cpu_flush_duration: Duration,
-}
-
-/// Perform one flush cycle: drain CPU profilers, drain the collector, write
-/// events to disk, and flush the writer. This is the only code path that
-/// touches EventWriter, and it runs exclusively on the flush thread.
-fn flush_once(
-    event_writer: &mut EventWriter,
-    shared: &SharedState,
-    drain_self: bool,
-) -> FlushStats {
-    let events_before = event_writer.events_written();
-    let cpu_events_time = std::time::Instant::now();
-    if shared.is_enabled() {
-        event_writer.flush_sources(shared);
-    }
-    let cpu_flush_duration = cpu_events_time.elapsed();
-
-    if drain_self {
-        // Periodically flush the flush thread's own TL buffer (queue samples + CPU events).
-        // We don't drain every cycle because each batch becomes its own trace segment;
-        // batching ~1s worth avoids writing tiny segments every 5ms.
-        buffer::drain_to_collector(&shared.collector);
-    }
-
-    let dropped = shared.collector.take_dropped_batches();
-    if dropped > 0 {
-        rate_limited!(Duration::from_secs(60), {
-            tracing::warn!(
-                dropped_batches = dropped,
-                "telemetry flush fell behind, dropped batches"
-            );
-        });
-    }
-
-    while let Some(batch) = shared.collector.next() {
-        if batch.event_count > 0
-            && let Err(e) = event_writer.write_encoded_batch(&batch)
-        {
-            rate_limited!(Duration::from_secs(60), {
-                tracing::warn!("failed to transcode batch: {e}");
-            });
-            shared.enabled.store(false, Ordering::Relaxed);
-            return FlushStats {
-                event_count: event_writer.events_written() - events_before,
-                dropped_batches: dropped as u64,
-                cpu_flush_duration,
-            };
-        }
-    }
-    if let Err(e) = event_writer.flush() {
-        rate_limited!(Duration::from_secs(60), {
-            tracing::warn!("failed to flush trace data: {e}");
-        });
-    }
-    FlushStats {
-        event_count: event_writer.events_written() - events_before,
-        dropped_batches: dropped as u64,
-        cpu_flush_duration,
-    }
 }
 
 /// Register telemetry callbacks on a runtime builder.
@@ -245,20 +155,14 @@ fn register_hooks(
             {
                 // Register as Blocking initially; worker threads will
                 // overwrite this to Worker(i) in resolve_worker_id.
-                // NOTE: `tokio::runtime::worker_index()` will always return `None` at this point
-                // so we can't utilize that here.
                 let tid = crate::telemetry::events::current_tid();
                 s_start
                     .thread_roles
                     .lock()
                     .unwrap()
                     .insert(tid, crate::telemetry::events::ThreadRole::Blocking);
-                // Sched event sampling is deferred to register_tid_if_needed(),
-                // which runs only for worker threads on their first poll/park.
-                // This avoids opening perf fds for blocking pool threads.
 
                 // Registers the current thread for the CPU-profiling fallback (ctimer).
-                // No-op when perf is the active backend (perf uses inherit).
                 let _ = dial9_perf_self_profile::register_current_thread();
             }
         })
@@ -339,1866 +243,13 @@ fn attach_runtime(
     Ok(runtime)
 }
 
-/// Cheap, cloneable handle for controlling telemetry from anywhere.
-///
-/// A handle may be in one of two modes:
-///
-/// - **Enabled** — backed by a real telemetry session; methods record
-///   events, control recording, and wrap spawned futures with wake
-///   tracking.
-/// - **Disabled** — an inert sentinel returned by
-///   [`TelemetryHandle::disabled`] and by [`TelemetryHandle::current`]
-///   when called from a thread that is not owned by a dial9 runtime.
-///   All methods are no-ops; [`spawn`](Self::spawn) falls back to
-///   [`tokio::spawn`] without wake tracking.
-///
-/// Use [`is_enabled`](Self::is_enabled) to distinguish the two modes.
-#[derive(Clone)]
-pub struct TelemetryHandle {
-    inner: Option<HandleInner>,
-}
-
-#[derive(Clone)]
-struct HandleInner {
-    shared: Arc<SharedState>,
-    control_tx: crate::primitives::sync::mpsc::SyncSender<ControlCommand>,
-}
-
-impl std::fmt::Debug for TelemetryHandle {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("TelemetryHandle")
-            .field("enabled", &self.is_enabled())
-            .finish_non_exhaustive()
-    }
-}
-
-impl TelemetryHandle {
-    pub(crate) fn enabled(
-        shared: Arc<SharedState>,
-        control_tx: crate::primitives::sync::mpsc::SyncSender<ControlCommand>,
-    ) -> Self {
-        Self {
-            inner: Some(HandleInner { shared, control_tx }),
-        }
-    }
-
-    /// Return an inert handle that is not connected to any telemetry
-    /// session. All methods are no-ops; [`spawn`](Self::spawn) falls
-    /// back to [`tokio::spawn`] without wake tracking.
-    pub fn disabled() -> Self {
-        Self { inner: None }
-    }
-
-    /// Whether this handle is connected to a live telemetry session.
-    ///
-    /// Returns `false` for handles obtained via
-    /// [`TelemetryHandle::disabled`], and for handles returned by
-    /// [`TelemetryHandle::current`] when called from a thread that is
-    /// not owned by a dial9 runtime.
-    pub fn is_enabled(&self) -> bool {
-        self.inner.is_some()
-    }
-
-    pub(crate) fn shared(&self) -> Option<&Arc<SharedState>> {
-        self.inner.as_ref().map(|i| &i.shared)
-    }
-
-    pub(crate) fn control_tx(
-        &self,
-    ) -> Option<&crate::primitives::sync::mpsc::SyncSender<ControlCommand>> {
-        self.inner.as_ref().map(|i| &i.control_tx)
-    }
-
-    /// Return the [`TelemetryHandle`] for the current thread.
-    ///
-    /// On threads owned by a dial9 runtime (workers and blocking
-    /// threads — installed via the runtime's `on_thread_start` hook,
-    /// cleared on `on_thread_stop`) this returns the live handle for
-    /// that runtime.
-    ///
-    /// On any other thread (including the caller of
-    /// `runtime.block_on(...)` on a `current_thread` runtime, threads
-    /// outside any tokio context, and threads owned by a runtime built
-    /// with telemetry disabled) this returns an inert handle whose
-    /// methods are all no-ops — see [`TelemetryHandle::disabled`].
-    ///
-    /// Use [`is_enabled`](Self::is_enabled) when you need to branch on
-    /// whether telemetry is actually live on the current thread.
-    pub fn current() -> Self {
-        CURRENT_HANDLE
-            .with(|cell| cell.borrow().clone())
-            .unwrap_or_else(Self::disabled)
-    }
-
-    /// Return the [`TelemetryHandle`] installed for the current thread,
-    /// or `None` if no dial9 runtime has claimed this thread.
-    ///
-    /// Prefer [`current`](Self::current) instead.
-    pub fn try_current() -> Option<Self> {
-        CURRENT_HANDLE.with(|cell| cell.borrow().clone())
-    }
-
-    /// Enable telemetry recording. No-op on a disabled handle.
-    pub fn enable(&self) {
-        if let Some(inner) = &self.inner {
-            inner.shared.enabled.store(true, Ordering::Relaxed);
-        }
-    }
-
-    /// Disable telemetry recording. No-op on a disabled handle.
-    pub fn disable(&self) {
-        if let Some(inner) = &self.inner {
-            inner.shared.enabled.store(false, Ordering::Relaxed);
-        }
-    }
-
-    /// Get a [`TracedHandle`](crate::traced::TracedHandle) for wrapping
-    /// futures with wake tracking, or `None` on a disabled handle.
-    pub(crate) fn traced_handle(&self) -> Option<crate::traced::TracedHandle> {
-        self.inner.as_ref().map(|i| crate::traced::TracedHandle {
-            shared: i.shared.clone(),
-        })
-    }
-
-    /// Record a user-defined [`Encodable`](crate::telemetry::buffer::Encodable) event.
-    ///
-    /// No-op on a disabled handle or when recording is paused.
-    pub(crate) fn record_encodable_event(&self, event: &dyn crate::telemetry::buffer::Encodable) {
-        if let Some(inner) = &self.inner {
-            inner
-                .shared
-                .if_enabled(|buf| buf.record_encodable_event(event));
-        }
-    }
-
-    /// Run a closure with direct access to the thread-local encoder.
-    ///
-    /// The closure is only invoked if telemetry is enabled.
-    /// No-op on a disabled handle or when recording is paused.
-    // TODO(GH-XXX): consider making this public as an alternative to record_event
-    // for zero-copy dynamic schema encoding
-    pub(crate) fn with_encoder(
-        &self,
-        f: impl FnOnce(&mut crate::telemetry::buffer::ThreadLocalEncoder<'_>),
-    ) {
-        if let Some(inner) = &self.inner {
-            inner.shared.if_enabled(|buf| buf.with_encoder(f));
-        }
-    }
-
-    /// Spawn a future on the ambient tokio runtime.
-    ///
-    /// On an enabled handle, the future is wrapped with wake-event
-    /// tracking. On a disabled handle, this is a passthrough to
-    /// [`tokio::spawn`].
-    ///
-    /// # Panics
-    ///
-    /// Panics if called from outside a tokio runtime context (same
-    /// as [`tokio::spawn`]).
-    #[track_caller]
-    pub fn spawn<F>(&self, future: F) -> tokio::task::JoinHandle<F::Output>
-    where
-        F: std::future::Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        match self.traced_handle() {
-            Some(traced_handle) => {
-                let _guard = InstrumentedSpawnGuard::enter();
-                tokio::spawn(crate::telemetry::TracedFuture::new(
-                    future,
-                    Some(traced_handle),
-                ))
-            }
-            None => tokio::spawn(future),
-        }
-    }
-
-    /// Spawn an instrumented future through a user-supplied spawn function.
-    ///
-    /// `spawn_fn` must synchronously perform a real Tokio spawn (or an
-    /// equivalent operation) before returning; do not defer the future or run
-    /// it with `block_on`. To record the resulting task as instrumented, spawn
-    /// on a dial9-traced runtime with task tracking enabled. The closure's
-    /// return value is forwarded back to the caller, so you can keep the
-    /// [`tokio::task::JoinHandle`], [`tokio::task::AbortHandle`], or whatever
-    /// the spawn function returns.
-    ///
-    /// # Examples
-    ///
-    /// Spawn into a [`tokio::task::JoinSet`]:
-    ///
-    /// ```rust,no_run
-    /// # use dial9_tokio_telemetry::telemetry::TelemetryHandle;
-    /// # use tokio::task::JoinSet;
-    /// # async fn work() {}
-    /// # async fn demo() {
-    /// let handle = TelemetryHandle::current();
-    /// let mut set: JoinSet<()> = JoinSet::new();
-    /// handle.spawn_with(work(), |f| set.spawn(f));
-    /// # }
-    /// ```
-    ///
-    /// Spawn into a [`tokio::task::JoinSet`] on a specific runtime:
-    ///
-    /// ```rust,no_run
-    /// # use dial9_tokio_telemetry::telemetry::TelemetryHandle;
-    /// # use tokio::runtime::Runtime;
-    /// # use tokio::task::JoinSet;
-    /// # async fn work() {}
-    /// # fn demo(runtime: &Runtime) {
-    /// let handle = TelemetryHandle::current();
-    /// let mut set: JoinSet<()> = JoinSet::new();
-    /// handle.spawn_with(work(), |f| set.spawn_on(f, runtime.handle()));
-    /// # }
-    /// ```
-    ///
-    /// [`TracedFuture<F>`]: crate::telemetry::TracedFuture
-    pub fn spawn_with<F, S>(
-        &self,
-        future: F,
-        spawn_fn: impl FnOnce(crate::telemetry::TracedFuture<F>) -> S,
-    ) -> S
-    where
-        F: std::future::Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        let traced_handle = self.traced_handle();
-        let future = crate::telemetry::TracedFuture::new(future, traced_handle.clone());
-        match traced_handle {
-            Some(_) => {
-                let _guard = InstrumentedSpawnGuard::enter();
-                spawn_fn(future)
-            }
-            None => spawn_fn(future),
-        }
-    }
-}
-
-/// Spawn a traced task on the current tokio runtime.
-///
-/// Like [`tokio::spawn`], but wraps the future with wake-event tracking
-/// when called from a thread owned by a dial9 runtime. On other threads,
-/// falls back to plain [`tokio::spawn`].
-///
-/// Equivalent to [`TelemetryHandle::current().spawn(future)`](TelemetryHandle::spawn).
-///
-/// # Panics
-///
-/// Panics if called from outside a tokio runtime context (same
-/// as [`tokio::spawn`]).
-#[track_caller]
-pub fn spawn<F>(future: F) -> tokio::task::JoinHandle<F::Output>
-where
-    F: std::future::Future + Send + 'static,
-    F::Output: Send + 'static,
-{
-    TelemetryHandle::current().spawn(future)
-}
-
-/// RAII guard that increments `INSTRUMENTED_SPAWN` on creation and
-/// decrements it on drop, even if the protected closure panics.
-struct InstrumentedSpawnGuard;
-
-impl InstrumentedSpawnGuard {
-    fn enter() -> Self {
-        INSTRUMENTED_SPAWN.with(|c| c.set(c.get().saturating_add(1)));
-        Self
-    }
-}
-
-impl Drop for InstrumentedSpawnGuard {
-    fn drop(&mut self) {
-        INSTRUMENTED_SPAWN.with(|c| c.set(c.get().saturating_sub(1)));
-    }
-}
-
-/// Handle for spawning instrumented futures on a specific runtime.
-///
-/// Returned by [`TraceRuntimeCoreBuilder::build`]. Unlike [`TelemetryHandle::spawn`]
-/// which uses `tokio::spawn()` (requiring an ambient runtime context), this type
-/// targets a specific runtime and works from any thread.
-///
-/// `Clone` is cheap — both inner handles are `Arc`-based.
-#[derive(Clone, Debug)]
-pub struct RuntimeTelemetryHandle {
-    runtime: tokio::runtime::Handle,
-    traced: Option<crate::traced::TracedHandle>,
-}
-
-impl RuntimeTelemetryHandle {
-    /// Spawn a future with wake-event tracking on this handle's runtime.
-    ///
-    /// On a handle obtained from a disabled [`TelemetryGuard`], wake
-    /// tracking is skipped and the future is spawned plainly.
-    #[track_caller]
-    pub fn spawn<F>(&self, future: F) -> tokio::task::JoinHandle<F::Output>
-    where
-        F: std::future::Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        match self.traced.clone() {
-            Some(traced_handle) => {
-                let _guard = InstrumentedSpawnGuard::enter();
-                self.runtime.spawn(crate::telemetry::TracedFuture::new(
-                    future,
-                    Some(traced_handle),
-                ))
-            }
-            None => self.runtime.spawn(future),
-        }
-    }
-
-    /// Spawn an instrumented future through a user-supplied spawn function.
-    ///
-    /// Mirrors [`TelemetryHandle::spawn_with`] for callers that already hold a
-    /// [`RuntimeTelemetryHandle`]. `spawn_fn` must synchronously perform a real
-    /// Tokio spawn (or an equivalent operation) before returning; do not defer
-    /// the future or run it with `block_on`. To record the resulting task as
-    /// instrumented, target a dial9-traced runtime with task tracking enabled,
-    /// typically via [`tokio::task::JoinSet::spawn_on`] with the appropriate
-    /// [`tokio::runtime::Handle`].
-    ///
-    /// # Examples
-    ///
-    /// Spawn into a [`tokio::task::JoinSet`] on a specific runtime:
-    ///
-    /// ```rust,no_run
-    /// # use dial9_tokio_telemetry::telemetry::RuntimeTelemetryHandle;
-    /// # use tokio::runtime::Runtime;
-    /// # use tokio::task::JoinSet;
-    /// # async fn work() {}
-    /// # fn demo(runtime: &Runtime, handle: RuntimeTelemetryHandle, set: &mut JoinSet<()>) {
-    /// handle.spawn_with(work(), |f| set.spawn_on(f, runtime.handle()));
-    /// # }
-    /// ```
-    ///
-    /// [`TracedFuture<F>`]: crate::telemetry::TracedFuture
-    pub fn spawn_with<F, S>(
-        &self,
-        future: F,
-        spawn_fn: impl FnOnce(crate::telemetry::TracedFuture<F>) -> S,
-    ) -> S
-    where
-        F: std::future::Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        let future = crate::telemetry::TracedFuture::new(future, self.traced.clone());
-        match self.traced {
-            Some(_) => {
-                let _guard = InstrumentedSpawnGuard::enter();
-                spawn_fn(future)
-            }
-            None => spawn_fn(future),
-        }
-    }
-}
-
-/// Holds the background worker thread and its stop signal.
-pub(crate) struct WorkerHandle {
-    shutdown: Option<tokio::sync::oneshot::Sender<Duration>>,
-    thread: Option<crate::primitives::thread::JoinHandle<()>>,
-}
-
-/// RAII guard returned by [`TracedRuntimeBuilder::build`].
-///
-/// A guard is always present on a [`TracedRuntime`], regardless of
-/// whether telemetry is enabled. When telemetry is disabled (because
-/// the user opted out via `enabled(false)` or because a lenient config
-/// path downgraded after a build failure), the guard is in an inert
-/// mode: all methods are no-ops, [`handle`](Self::handle) returns an
-/// inert [`TelemetryHandle`], and [`graceful_shutdown`](Self::graceful_shutdown)
-/// is a successful no-op.
-///
-/// Use [`is_enabled`](Self::is_enabled) to distinguish the two modes.
-pub struct TelemetryGuard {
-    inner: GuardInner,
-}
-
-enum GuardInner {
-    Enabled(EnabledGuard),
-    Disabled,
-}
-
-struct EnabledGuard {
-    handle: TelemetryHandle,
-    flush_thread: Option<crate::primitives::thread::JoinHandle<()>>,
-    worker: Option<WorkerHandle>,
-}
-
-impl std::fmt::Debug for TelemetryGuard {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("TelemetryGuard")
-            .field("enabled", &self.is_enabled())
-            .finish_non_exhaustive()
-    }
-}
-
-impl TelemetryGuard {
-    pub(crate) fn enabled(
-        handle: TelemetryHandle,
-        flush_thread: Option<crate::primitives::thread::JoinHandle<()>>,
-        worker: Option<WorkerHandle>,
-    ) -> Self {
-        Self {
-            inner: GuardInner::Enabled(EnabledGuard {
-                handle,
-                flush_thread,
-                worker,
-            }),
-        }
-    }
-
-    pub(crate) fn disabled() -> Self {
-        Self {
-            inner: GuardInner::Disabled,
-        }
-    }
-
-    /// Whether this guard owns a live telemetry session.
-    ///
-    /// Returns `false` for guards created by `enabled(false)` configs
-    /// or by lenient configs that downgraded after a build failure.
-    pub fn is_enabled(&self) -> bool {
-        matches!(self.inner, GuardInner::Enabled(_))
-    }
-
-    /// Get a cloneable handle for controlling telemetry.
-    ///
-    /// On a disabled guard this returns an inert handle whose methods
-    /// are all no-ops — see [`TelemetryHandle::disabled`].
-    pub fn handle(&self) -> TelemetryHandle {
-        match &self.inner {
-            GuardInner::Enabled(eg) => eg.handle.clone(),
-            GuardInner::Disabled => TelemetryHandle::disabled(),
-        }
-    }
-
-    /// Monotonic start time of the telemetry session in nanoseconds, if
-    /// telemetry is enabled.
-    pub fn start_time(&self) -> Option<u64> {
-        self.shared().map(|s| s.start_time_ns)
-    }
-
-    /// Enable telemetry recording. No-op on a disabled guard.
-    pub fn enable(&self) {
-        if let GuardInner::Enabled(eg) = &self.inner {
-            eg.handle.enable();
-        }
-    }
-
-    /// Disable telemetry recording. No-op on a disabled guard.
-    pub fn disable(&self) {
-        if let GuardInner::Enabled(eg) = &self.inner {
-            eg.handle.disable();
-        }
-    }
-
-    /// Access the shared state for reuse by additional runtimes.
-    pub(crate) fn shared(&self) -> Option<&Arc<SharedState>> {
-        match &self.inner {
-            GuardInner::Enabled(eg) => eg.handle.shared(),
-            GuardInner::Disabled => None,
-        }
-    }
-
-    pub(crate) fn control_tx(
-        &self,
-    ) -> Option<&crate::primitives::sync::mpsc::SyncSender<ControlCommand>> {
-        match &self.inner {
-            GuardInner::Enabled(eg) => eg.handle.control_tx(),
-            GuardInner::Disabled => None,
-        }
-    }
-
-    /// Attach a tokio runtime to this telemetry session.
-    ///
-    /// Returns a builder that lets you configure per-runtime settings
-    /// (e.g. task tracking) before building the runtime.
-    ///
-    /// On a disabled guard the resulting builder produces a plain tokio
-    /// runtime with no telemetry hooks installed.
-    ///
-    /// ```rust,no_run
-    /// # use dial9_tokio_telemetry::telemetry::{NullWriter, TelemetryCore};
-    /// # fn main() -> std::io::Result<()> {
-    /// let guard = TelemetryCore::builder()
-    ///     .writer(NullWriter)
-    ///     .build()?;
-    /// guard.enable();
-    ///
-    /// let mut builder = tokio::runtime::Builder::new_multi_thread();
-    /// builder.worker_threads(4).enable_all();
-    /// let (runtime, handle) = guard.trace_runtime("main").build(builder)?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn trace_runtime(&self, name: impl Into<String>) -> TraceRuntimeCoreBuilder<'_> {
-        TraceRuntimeCoreBuilder {
-            guard: self,
-            name: name.into(),
-            task_tracking: false,
-        }
-    }
-
-    /// Send FinalizeAndStop to the flush thread, join it, then drain the
-    /// caller's thread-local buffer into the collector so the flush thread
-    /// picks up any stragglers. No-op when telemetry is disabled.
-    fn stop_flush_thread(&mut self) {
-        let GuardInner::Enabled(eg) = &mut self.inner else {
-            return;
-        };
-        // Drain the current thread's buffer (e.g. main thread in block_on)
-        // which may contain TaskSpawn events that were never flushed.
-        if let Some(shared) = eg.handle.shared() {
-            buffer::drain_to_collector(&shared.collector);
-        }
-
-        // Tell the flush thread to do a final flush + finalize, then exit.
-        let (ack_tx, ack_rx) = crate::primitives::sync::mpsc::sync_channel(0);
-        if let Some(tx) = eg.handle.control_tx()
-            && tx.send(ControlCommand::FinalizeAndStop(ack_tx)).is_ok()
-        {
-            let _ = ack_rx.recv();
-        }
-        if let Some(t) = eg.flush_thread.take() {
-            let _ = t.join();
-        }
-    }
-
-    /// Flush remaining events, seal the final segment, and wait for the
-    /// background worker to drain (symbolize, compress, upload to S3).
-    ///
-    /// **Call this after the runtime has been dropped** so that Tokio worker
-    /// threads have exited and their thread-local telemetry buffers have been
-    /// flushed to the central collector.
-    ///
-    /// On a disabled guard this is a successful no-op — there is no
-    /// flush thread or background worker to drain.
-    ///
-    /// ```rust,no_run
-    /// # use dial9_tokio_telemetry::telemetry::{RotatingWriter, TracedRuntime};
-    /// # use std::time::Duration;
-    /// # fn main() -> std::io::Result<()> {
-    /// # let writer = RotatingWriter::new("/tmp/t.bin", 1024, 4096)?;
-    /// # let builder = tokio::runtime::Builder::new_multi_thread();
-    /// let (runtime, guard) = TracedRuntime::build_and_start(builder, writer)?;
-    /// runtime.block_on(async { /* ... */ });
-    /// drop(runtime); // worker threads exit, flushing thread-local buffers
-    /// guard.graceful_shutdown(Duration::from_secs(5))?;
-    /// # Ok(())
-    /// # }
-    /// ```
-    ///
-    /// Consumes the guard so `Drop` becomes a no-op.
-    pub fn graceful_shutdown(mut self, timeout: Duration) -> Result<(), std::io::Error> {
-        tracing::debug!(target: "dial9_telemetry", "graceful_shutdown starting");
-
-        // 1. Stop flush thread (flushes + finalizes the last segment).
-        // No-op when disabled.
-        self.stop_flush_thread();
-        tracing::debug!(target: "dial9_telemetry", "flush thread joined, segment sealed");
-
-        // 2. Signal worker to drain with the given timeout and wait
-        if let GuardInner::Enabled(eg) = &mut self.inner
-            && let Some(ref mut w) = eg.worker
-        {
-            tracing::debug!(target: "dial9_telemetry", timeout_secs = timeout.as_secs(), "waiting for worker drain");
-            if let Some(tx) = w.shutdown.take() {
-                let _ = tx.send(timeout);
-            }
-            if let Some(t) = w.thread.take()
-                && let Err(e) = t.join()
-            {
-                tracing::error!(target: "dial9_telemetry", panic = ?e, "worker thread panicked during shutdown");
-            }
-            tracing::debug!(target: "dial9_telemetry", "worker finished");
-        }
-
-        Ok(())
-    }
-}
-
-impl Drop for TelemetryGuard {
-    fn drop(&mut self) {
-        // 1. Stop the flush thread (flushes + finalizes). No-op when disabled.
-        self.stop_flush_thread();
-
-        // 2. Hard shutdown: drop the sender without sending — worker sees
-        // RecvError and exits without draining. No need to join the thread.
-        // For graceful drain, use graceful_shutdown() instead.
-        if let GuardInner::Enabled(eg) = &mut self.inner
-            && let Some(ref mut w) = eg.worker
-        {
-            w.shutdown.take();
-        }
-    }
-}
-
-/// Marker: no trace path has been set yet.
-#[derive(Debug)]
-#[non_exhaustive]
-pub struct NoTracePath;
-/// Marker: a trace path has been set.
-#[derive(Debug)]
-#[non_exhaustive]
-pub struct HasTracePath;
-
-/// Marker: no pipeline strategy has been chosen yet. From this state the
-/// builder can transition to either S3 (via `with_s3_uploader`) or a custom
-/// pipeline (via `with_custom_pipeline`).
-#[derive(Debug)]
-#[non_exhaustive]
-pub struct PipelineUnset;
-
-/// Marker: the S3 preset has been selected. `with_s3_client` is available
-/// to bind a pre-built client; `with_custom_pipeline` is not in scope.
-#[derive(Debug)]
-#[non_exhaustive]
-pub struct PipelineS3;
-
-/// Marker: a custom pipeline has been configured. No further pipeline
-/// methods are available.
-#[derive(Debug)]
-#[non_exhaustive]
-pub struct PipelineCustom;
-
-enum PipelineConfig {
-    Unset,
-    #[cfg(feature = "worker-s3")]
-    S3(crate::background_task::S3PipelineUploader),
-    Custom(Vec<Box<dyn crate::background_task::SegmentProcessor>>),
-}
-
-/// Builder for configuring a traced Tokio runtime.
-pub struct TracedRuntimeBuilder<P = NoTracePath, M = PipelineUnset> {
-    enabled: bool,
-    task_tracking_enabled: bool,
-    task_dump_config: Option<crate::telemetry::task_dump_config::TaskDumpConfig>,
-    trace_path: Option<PathBuf>,
-    runtime_name: Option<String>,
-    #[cfg(feature = "cpu-profiling")]
-    cpu_profiling_config: Option<crate::telemetry::cpu_profile::CpuProfilingConfig>,
-    #[cfg(feature = "cpu-profiling")]
-    sched_event_config: Option<crate::telemetry::cpu_profile::SchedEventConfig>,
-    pipeline: PipelineConfig,
-    /// Static segment metadata to inject into every rotated segment's
-    /// header. The S3 preset populates this from `S3Config::as_metadata`
-    /// so traces stay self-describing.
-    segment_metadata: Vec<(String, String)>,
-    worker_poll_interval: Option<Duration>,
-    worker_metrics_sink: Option<metrique_writer::BoxEntrySink>,
-    _marker: std::marker::PhantomData<(P, M)>,
-}
-
-impl<P, M> std::fmt::Debug for TracedRuntimeBuilder<P, M> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("TracedRuntimeBuilder")
-            .finish_non_exhaustive()
-    }
-}
-
-// Methods available regardless of trace-path or pipeline state.
-impl<P, M> TracedRuntimeBuilder<P, M> {
-    /// Set to `false` to build a plain runtime with no telemetry
-    /// installed and a dummy [`TelemetryGuard`]. Defaults to `true`.
-    ///
-    /// Unlike [`TelemetryGuard::enable`]/[`TelemetryGuard::disable`]
-    /// (which toggle recording at runtime), this controls whether
-    /// telemetry hooks and threads are installed at all.
-    pub fn install(mut self, enabled: bool) -> Self {
-        self.enabled = enabled;
-        self
-    }
-
-    /// Enable or disable task spawn/terminate tracking.
-    pub fn with_task_tracking(mut self, enabled: bool) -> Self {
-        self.task_tracking_enabled = enabled;
-        self
-    }
-
-    /// Capture async backtraces at yield points for tasks that stay idle
-    /// longer than the configured threshold.
-    ///
-    /// Requires the `taskdump` crate feature to actually record events
-    pub fn with_task_dumps(
-        mut self,
-        config: crate::telemetry::task_dump_config::TaskDumpConfig,
-    ) -> Self {
-        if cfg!(not(feature = "taskdump")) {
-            tracing::warn!(
-                "taskdumps enabled but `taskdump` feature was not. No task dumps will be captured."
-            )
-        }
-        self.task_dump_config = Some(config);
-        self
-    }
-
-    /// Set a human-readable name for this runtime. Used in segment metadata
-    /// to map runtime indices to names for the trace viewer.
-    pub fn with_runtime_name(mut self, name: impl Into<String>) -> Self {
-        self.runtime_name = Some(name.into());
-        self
-    }
-
-    /// Set static metadata embedded as a `SegmentMetadata` event in every
-    /// sealed segment file. Read back during analysis and attached to every
-    /// Span.
-    ///
-    /// [`with_s3_uploader`](Self::with_s3_uploader) injects bucket /
-    /// service_name / instance_path / boot_id automatically; call this
-    /// method when using [`with_custom_pipeline`](Self::with_custom_pipeline)
-    /// (or no pipeline) and you still want those entries — or when you want
-    /// to override the preset's defaults.
-    ///
-    /// Repeated calls **replace** the metadata, matching how
-    /// `with_s3_uploader` overwrites on a second call. The last call wins,
-    /// so `with_segment_metadata` placed *after* `with_s3_uploader`
-    /// overrides the preset's injection.
-    pub fn with_segment_metadata(mut self, entries: Vec<(String, String)>) -> Self {
-        self.segment_metadata = entries;
-        self
-    }
-
-    /// Enable CPU profiling with the given configuration (Linux only).
-    #[cfg(feature = "cpu-profiling")]
-    pub fn with_cpu_profiling(
-        mut self,
-        config: crate::telemetry::cpu_profile::CpuProfilingConfig,
-    ) -> Self {
-        self.cpu_profiling_config = Some(config);
-        self
-    }
-
-    /// Enable per-worker scheduler event capture (Linux only).
-    #[cfg(feature = "cpu-profiling")]
-    pub fn with_sched_events(
-        mut self,
-        config: crate::telemetry::cpu_profile::SchedEventConfig,
-    ) -> Self {
-        self.sched_event_config = Some(config);
-        self
-    }
-
-    /// Set how often the background worker polls for sealed segments.
-    pub fn with_worker_poll_interval(mut self, interval: Duration) -> Self {
-        self.worker_poll_interval = Some(interval);
-        self
-    }
-
-    /// Set a metrics sink for the background worker.
-    pub fn with_worker_metrics_sink(mut self, sink: metrique_writer::BoxEntrySink) -> Self {
-        self.worker_metrics_sink = Some(sink);
-        self
-    }
-
-    /// Attach a new runtime to an existing telemetry session.
-    ///
-    /// This reuses the `SharedState`, flush thread, writer, and CPU profiler
-    /// from the original `TelemetryGuard`. Only the tokio callbacks are
-    /// registered on the new builder. The new runtime's workers get a unique
-    /// runtime index so their `WorkerId`s don't collide with existing runtimes.
-    pub fn build_and_attach_to_telemetry(
-        self,
-        mut builder: tokio::runtime::Builder,
-        guard: &TelemetryGuard,
-    ) -> std::io::Result<tokio::runtime::Runtime> {
-        let (Some(shared), Some(control_tx)) = (guard.shared(), guard.control_tx()) else {
-            // Disabled guard: produce a plain tokio runtime with no
-            // telemetry hooks so attaching still works gracefully.
-            return builder.build();
-        };
-        attach_runtime(
-            shared,
-            builder,
-            self.runtime_name,
-            control_tx,
-            self.task_tracking_enabled,
-        )
-    }
-
-    fn into_state<Q, N>(self) -> TracedRuntimeBuilder<Q, N> {
-        TracedRuntimeBuilder {
-            enabled: self.enabled,
-            task_tracking_enabled: self.task_tracking_enabled,
-            task_dump_config: self.task_dump_config,
-            trace_path: self.trace_path,
-            runtime_name: self.runtime_name,
-            #[cfg(feature = "cpu-profiling")]
-            cpu_profiling_config: self.cpu_profiling_config,
-            #[cfg(feature = "cpu-profiling")]
-            sched_event_config: self.sched_event_config,
-            pipeline: self.pipeline,
-            segment_metadata: self.segment_metadata,
-            worker_poll_interval: self.worker_poll_interval,
-            worker_metrics_sink: self.worker_metrics_sink,
-            _marker: std::marker::PhantomData,
-        }
-    }
-}
-
-// Pipeline-strategy entry points: only available before a strategy has
-// been chosen, so the user picks S3 OR a custom pipeline, not both.
-impl<P> TracedRuntimeBuilder<P, PipelineUnset> {
-    /// Configure the S3 upload preset for sealed trace segments.
-    ///
-    /// The resulting pipeline is `[Gzip, S3]` (with `[Symbolize, ...]`
-    /// prepended when CPU profiling is enabled). After this call, only
-    /// [`with_s3_client`](TracedRuntimeBuilder::with_s3_client) and a
-    /// repeated [`with_s3_uploader`](TracedRuntimeBuilder::with_s3_uploader)
-    /// override are available — `with_custom_pipeline` is no longer in scope.
-    #[cfg(feature = "worker-s3")]
-    pub fn with_s3_uploader(
-        mut self,
-        config: crate::background_task::s3::S3Config,
-    ) -> TracedRuntimeBuilder<P, PipelineS3> {
-        self.segment_metadata = config
-            .as_metadata()
-            .map(|(k, v)| (k.to_string(), v.to_string()))
-            .collect();
-        self.pipeline = PipelineConfig::S3(crate::background_task::S3PipelineUploader::new(
-            config, None,
-        ));
-        self.into_state()
-    }
-
-    /// Configure a fully custom processor pipeline. The closure receives a
-    /// [`PipelineBuilder`](crate::background_task::PipelineBuilder); chain
-    /// methods like `.gzip()`, `.write_back()`, `.s3(cfg)` for built-ins
-    /// and `.pipe(processor)` for user-supplied processors.
-    ///
-    /// Mutually exclusive with [`with_s3_uploader`](Self::with_s3_uploader).
-    ///
-    /// This is the "full control" path: the resulting pipeline is exactly
-    /// what the closure builds, with nothing prepended or appended. In
-    /// particular, unlike the S3 preset, this path does **not**:
-    /// - auto-populate writer-side segment metadata — call
-    ///   [`with_segment_metadata`](Self::with_segment_metadata) if you want
-    ///   identity entries (service, host, etc.) embedded in trace files.
-    /// - auto-prepend the `Symbolize` step when CPU profiling is enabled.
-    ///   Chain
-    ///   [`.symbolize()`](crate::background_task::PipelineBuilder::symbolize)
-    ///   first if you want symbolized stack frames.
-    pub fn with_custom_pipeline<F>(mut self, build: F) -> TracedRuntimeBuilder<P, PipelineCustom>
-    where
-        F: FnOnce(
-            crate::background_task::PipelineBuilder,
-        ) -> crate::background_task::PipelineBuilder,
-    {
-        let pipeline = build(crate::background_task::PipelineBuilder::new());
-        self.pipeline = PipelineConfig::Custom(pipeline.into_processors());
-        self.into_state()
-    }
-}
-
-// S3 mode — once the S3 preset is chosen, only S3-specific tweaks remain.
-#[cfg(feature = "worker-s3")]
-impl<P> TracedRuntimeBuilder<P, PipelineS3> {
-    /// Provide a pre-built S3 client (for custom credentials or endpoints).
-    /// Replaces any client previously bound to the configured S3 uploader.
-    pub fn with_s3_client(mut self, client: aws_sdk_s3::Client) -> Self {
-        if let PipelineConfig::S3(ref mut uploader) = self.pipeline {
-            uploader.set_client(client);
-        }
-        self
-    }
-
-    /// Replace the configured S3 uploader. A client previously bound via
-    /// [`with_s3_client`](Self::with_s3_client) is carried over to the new
-    /// uploader so that call order between the two is irrelevant.
-    pub fn with_s3_uploader(mut self, config: crate::background_task::s3::S3Config) -> Self {
-        self.segment_metadata = config
-            .as_metadata()
-            .map(|(k, v)| (k.to_string(), v.to_string()))
-            .collect();
-        let carried = match &mut self.pipeline {
-            PipelineConfig::S3(uploader) => uploader.take_client(),
-            _ => None,
-        };
-        self.pipeline = PipelineConfig::S3(crate::background_task::S3PipelineUploader::new(
-            config, carried,
-        ));
-        self
-    }
-}
-
-impl<M> TracedRuntimeBuilder<NoTracePath, M> {
-    /// Set the trace output path. This transitions the builder to
-    /// `HasTracePath`, enabling `build()` and `build_and_start()`.
-    pub fn with_trace_path(
-        mut self,
-        path: impl Into<PathBuf>,
-    ) -> TracedRuntimeBuilder<HasTracePath, M> {
-        self.trace_path = Some(path.into());
-        self.into_state()
-    }
-
-    /// Build with a custom writer (for tests or `NullWriter`).
-    /// No background worker is spawned.
-    pub fn build_with_writer(
-        self,
-        builder: tokio::runtime::Builder,
-        writer: impl TraceWriter + 'static,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        self.into_state::<HasTracePath, M>()
-            .build_inner(builder, Box::new(writer))
-    }
-
-    /// Build with a custom writer and immediately enable recording.
-    pub fn build_and_start_with_writer(
-        self,
-        builder: tokio::runtime::Builder,
-        writer: impl TraceWriter + 'static,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        let (runtime, guard) = self.build_with_writer(builder, writer)?;
-        guard.enable();
-        Ok((runtime, guard))
-    }
-
-    /// Build the traced runtime. No background worker is spawned
-    /// (use `with_trace_path()` first for worker support).
-    pub fn build(
-        self,
-        builder: tokio::runtime::Builder,
-        writer: impl TraceWriter + 'static,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        self.build_with_writer(builder, writer)
-    }
-
-    /// Build and immediately enable recording.
-    pub fn build_and_start(
-        self,
-        builder: tokio::runtime::Builder,
-        writer: impl TraceWriter + 'static,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        self.build_and_start_with_writer(builder, writer)
-    }
-}
-
-impl<M> TracedRuntimeBuilder<HasTracePath, M> {
-    /// Set the trace output path (no-op, already set).
-    pub fn with_trace_path(mut self, path: impl Into<PathBuf>) -> Self {
-        self.trace_path = Some(path.into());
-        self
-    }
-
-    /// Build the traced runtime with a `RotatingWriter`.
-    ///
-    /// The background worker is auto-spawned when cpu-profiling or any
-    /// pipeline strategy is configured. Recording starts disabled; call
-    /// [`TelemetryGuard::enable`] to begin, or use
-    /// [`build_and_start`](Self::build_and_start).
-    pub fn build(
-        self,
-        builder: tokio::runtime::Builder,
-        writer: RotatingWriter,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        self.build_inner(builder, Box::new(writer))
-    }
-
-    /// Build the traced runtime and immediately enable recording.
-    pub fn build_and_start(
-        self,
-        builder: tokio::runtime::Builder,
-        writer: RotatingWriter,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        let (runtime, guard) = self.build(builder, writer)?;
-        guard.enable();
-        Ok((runtime, guard))
-    }
-
-    /// Build with a custom writer (for tests). The background worker is
-    /// still spawned if cpu-profiling or any pipeline strategy is configured
-    /// and `trace_path` is set.
-    pub fn build_with_writer(
-        self,
-        builder: tokio::runtime::Builder,
-        writer: impl TraceWriter + 'static,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        self.build_inner(builder, Box::new(writer))
-    }
-
-    /// Build with a custom writer and immediately enable recording.
-    pub fn build_and_start_with_writer(
-        self,
-        builder: tokio::runtime::Builder,
-        writer: impl TraceWriter + 'static,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        let (runtime, guard) = self.build_with_writer(builder, writer)?;
-        guard.enable();
-        Ok((runtime, guard))
-    }
-
-    fn build_inner(
-        self,
-        builder: tokio::runtime::Builder,
-        writer: Box<dyn TraceWriter>,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        if !self.enabled {
-            return TracedRuntime::build_disabled(builder);
-        }
-
-        let processors = assemble_processors(
-            #[cfg(feature = "cpu-profiling")]
-            self.cpu_profiling_config.is_some(),
-            self.pipeline,
-        );
-
-        let core_builder = TelemetryCore::builder()
-            .writer(writer)
-            .maybe_trace_path(self.trace_path)
-            .maybe_task_dump_config(self.task_dump_config)
-            .maybe_worker_poll_interval(self.worker_poll_interval)
-            .maybe_worker_metrics_sink(self.worker_metrics_sink)
-            .processors(processors)
-            .segment_metadata(self.segment_metadata);
-
-        #[cfg(feature = "cpu-profiling")]
-        let core_builder = core_builder
-            .maybe_cpu_profiling(self.cpu_profiling_config)
-            .maybe_sched_events(self.sched_event_config);
-
-        let guard = core_builder.build()?;
-        let control_tx = guard
-            .control_tx()
-            .expect("TelemetryCore::builder().build() always returns an enabled guard")
-            .clone();
-        let shared = guard
-            .shared()
-            .expect("TelemetryCore::builder().build() always returns an enabled guard");
-        let runtime = attach_runtime(
-            shared,
-            builder,
-            self.runtime_name,
-            &control_tx,
-            self.task_tracking_enabled,
-        )?;
-        Ok((runtime, guard))
-    }
-}
-
-/// Build the final processor pipeline.
-///
-/// `Symbolize` is auto-prepended for the built-in presets (`Unset`, `S3`)
-/// when CPU profiling is enabled. The `Custom` path is "full control" — the
-/// user's processor list is passed through verbatim, and they're expected to
-/// chain [`PipelineBuilder::symbolize`](crate::background_task::PipelineBuilder::symbolize)
-/// themselves if they want symbolization.
-///
-/// Behaviour matrix:
-///
-/// | strategy | CPU profiling on               | CPU profiling off |
-/// |----------|--------------------------------|-------------------|
-/// | Unset    | `[Symbolize, Gzip, WriteBack]` | (worker skipped)  |
-/// | S3       | `[Symbolize, Gzip, S3]`        | `[Gzip, S3]`      |
-/// | Custom   | `[...user]`                    | `[...user]`       |
-fn assemble_processors(
-    #[cfg(feature = "cpu-profiling")] cpu_profiling_enabled: bool,
-    pipeline: PipelineConfig,
-) -> Vec<Box<dyn crate::background_task::SegmentProcessor>> {
-    #[cfg(not(feature = "cpu-profiling"))]
-    let cpu_profiling_enabled = false;
-
-    if matches!(pipeline, PipelineConfig::Unset) && !cpu_profiling_enabled {
-        return Vec::new();
-    }
-
-    let mut processors: Vec<Box<dyn crate::background_task::SegmentProcessor>> = Vec::new();
-    match pipeline {
-        PipelineConfig::Unset => {
-            #[cfg(feature = "cpu-profiling")]
-            if cpu_profiling_enabled {
-                processors.push(Box::new(crate::background_task::SymbolizeProcessor));
-            }
-            processors.push(Box::new(crate::background_task::GzipCompressor));
-            processors.push(Box::new(crate::background_task::WriteBackProcessor));
-        }
-        #[cfg(feature = "worker-s3")]
-        PipelineConfig::S3(uploader) => {
-            #[cfg(feature = "cpu-profiling")]
-            if cpu_profiling_enabled {
-                processors.push(Box::new(crate::background_task::SymbolizeProcessor));
-            }
-            processors.push(Box::new(crate::background_task::GzipCompressor));
-            processors.push(Box::new(uploader));
-        }
-        PipelineConfig::Custom(user) => {
-            processors.extend(user);
-        }
-    }
-    processors
-}
-
-/// Builder for attaching a runtime to an existing telemetry session.
-///
-/// Created by [`TelemetryGuard::trace_runtime`]. Call [`.build()`](Self::build)
-/// with a [`tokio::runtime::Builder`] to install hooks and build the runtime.
-#[must_use]
-#[derive(Debug)]
-pub struct TraceRuntimeCoreBuilder<'a> {
-    guard: &'a TelemetryGuard,
-    name: String,
-    task_tracking: bool,
-}
-
-impl<'a> TraceRuntimeCoreBuilder<'a> {
-    /// Enable or disable task spawn/terminate tracking for this runtime.
-    /// Defaults to `false`.
-    pub fn task_tracking(mut self, enabled: bool) -> Self {
-        self.task_tracking = enabled;
-        self
-    }
-
-    /// Install telemetry hooks, build the runtime, and reserve worker IDs.
-    ///
-    /// Returns the runtime and a [`RuntimeTelemetryHandle`] for spawning
-    /// instrumented futures via [`RuntimeTelemetryHandle::spawn`].
-    pub fn build(
-        self,
-        mut builder: tokio::runtime::Builder,
-    ) -> std::io::Result<(tokio::runtime::Runtime, RuntimeTelemetryHandle)> {
-        let (Some(shared), Some(control_tx), Some(traced)) = (
-            self.guard.shared(),
-            self.guard.control_tx(),
-            self.guard.handle().traced_handle(),
-        ) else {
-            // Disabled guard: build a plain tokio runtime and return a
-            // RuntimeTelemetryHandle that effectively short-circuits to
-            // tokio::spawn.
-            let runtime = builder.build()?;
-            let handle = RuntimeTelemetryHandle {
-                runtime: runtime.handle().clone(),
-                traced: None,
-            };
-            return Ok((runtime, handle));
-        };
-        let runtime = attach_runtime(
-            shared,
-            builder,
-            Some(self.name),
-            control_tx,
-            self.task_tracking,
-        )?;
-        let handle = RuntimeTelemetryHandle {
-            runtime: runtime.handle().clone(),
-            traced: Some(traced),
-        };
-        Ok((runtime, handle))
-    }
-}
-
-/// Entry point for creating a telemetry session decoupled from any tokio runtime.
-///
-/// Use [`TelemetryCore::builder()`] to configure the session, then call
-/// [`TelemetryGuard::trace_runtime`] to attach one or more runtimes.
-///
-/// ```rust,no_run
-/// # use dial9_tokio_telemetry::telemetry::{RotatingWriter, TelemetryCore};
-/// # fn main() -> std::io::Result<()> {
-/// let writer = RotatingWriter::single_file("/tmp/trace.bin")?;
-/// let guard = TelemetryCore::builder()
-///     .writer(writer)
-///     .build()?;
-/// guard.enable();
-///
-/// let mut builder = tokio::runtime::Builder::new_multi_thread();
-/// builder.worker_threads(4).enable_all();
-/// let (runtime, handle) = guard.trace_runtime("main").build(builder)?;
-/// # Ok(())
-/// # }
-/// ```
-#[derive(Debug)]
-pub struct TelemetryCore;
-
-#[bon::bon]
-impl TelemetryCore {
-    /// Build a telemetry session. Recording starts disabled; call
-    /// [`TelemetryGuard::enable`] to begin recording.
-    #[builder(state_mod = telemetry_core_builder)]
-    pub fn new(
-        /// The pipeline of [`SegmentProcessor`](crate::background_task::SegmentProcessor)s
-        /// to run on each sealed segment. When empty the background worker
-        /// is not spawned.
-        #[builder(field)]
-        processors: Vec<Box<dyn crate::background_task::SegmentProcessor>>,
-        /// Static segment metadata injected into every rotated segment's
-        /// header. Empty by default; the S3 preset populates it from the
-        /// configured `S3Config` so traces stay self-describing.
-        #[builder(field)]
-        segment_metadata: Vec<(String, String)>,
-        /// S3 upload configuration.
-        #[cfg(feature = "worker-s3")]
-        #[builder(field)]
-        s3_config: Option<crate::background_task::s3::S3Config>,
-        /// Pre-built S3 client.
-        #[cfg(feature = "worker-s3")]
-        #[builder(field)]
-        s3_client: Option<aws_sdk_s3::Client>,
-        /// The trace writer (e.g. [`RotatingWriter`], [`NullWriter`]).
-        writer: impl TraceWriter + 'static,
-        /// Path for trace output. Enables the background worker when any
-        /// segment processors are configured.
-        #[builder(into)]
-        trace_path: Option<PathBuf>,
-        /// Capture async backtraces at yield points. Requires the `taskdump`
-        /// crate feature to actually record events.
-        task_dump_config: Option<crate::telemetry::task_dump_config::TaskDumpConfig>,
-        /// Enable CPU profiling (Linux only).
-        #[cfg(feature = "cpu-profiling")]
-        cpu_profiling: Option<crate::telemetry::cpu_profile::CpuProfilingConfig>,
-        /// Enable scheduler event capture (Linux only).
-        #[cfg(feature = "cpu-profiling")]
-        sched_events: Option<crate::telemetry::cpu_profile::SchedEventConfig>,
-        /// How often the background worker polls for sealed segments.
-        worker_poll_interval: Option<Duration>,
-        /// Metrics sink for the flush/worker threads.
-        worker_metrics_sink: Option<metrique_writer::BoxEntrySink>,
-    ) -> std::io::Result<TelemetryGuard> {
-        let start_mono_ns = crate::telemetry::events::clock_monotonic_ns();
-        let rng_seed = task_dump_config.as_ref().and_then(|cfg| cfg.rng_seed());
-        let shared = Arc::new(SharedState::new(start_mono_ns, rng_seed));
-        if let Some(cfg) = task_dump_config.as_ref() {
-            shared.task_dumps_enabled.store(true, Ordering::Relaxed);
-            shared
-                .task_dump_idle_threshold_ns
-                .store(cfg.idle_threshold().as_nanos() as u64, Ordering::Relaxed);
-        }
-
-        // Determine the pipeline strategy from the builder fields, then
-        // delegate to `assemble_processors` — the single source of truth for
-        // which processors are used in each configuration.
-        #[allow(unused_mut)]
-        let mut segment_metadata = segment_metadata;
-
-        #[allow(unused_variables)]
-        let pipeline = if !processors.is_empty() {
-            PipelineConfig::Custom(processors)
-        } else {
-            #[cfg(feature = "worker-s3")]
-            if let Some(config) = s3_config {
-                if segment_metadata.is_empty() {
-                    segment_metadata = config
-                        .as_metadata()
-                        .map(|(k, v)| (k.to_string(), v.to_string()))
-                        .collect();
-                }
-                PipelineConfig::S3(crate::background_task::S3PipelineUploader::new(
-                    config, s3_client,
-                ))
-            } else {
-                PipelineConfig::Unset
-            }
-            #[cfg(not(feature = "worker-s3"))]
-            PipelineConfig::Unset
-        };
-
-        let processors = assemble_processors(
-            #[cfg(feature = "cpu-profiling")]
-            cpu_profiling.is_some(),
-            pipeline,
-        );
-
-        #[allow(unused_mut)]
-        let mut event_writer = EventWriter::new(Box::new(writer));
-
-        if !segment_metadata.is_empty() {
-            event_writer.update_segment_metadata(segment_metadata);
-        }
-
-        #[cfg(feature = "cpu-profiling")]
-        {
-            if let Some(ref config) = cpu_profiling {
-                match crate::telemetry::cpu_profile::CpuProfiler::start(config.clone()) {
-                    Ok(sampler) => shared.push_source(Box::new(sampler)),
-                    Err(e) => rate_limited!(Duration::from_secs(60), {
-                        tracing::warn!("failed to start CPU profiler: {e}");
-                    }),
-                }
-            }
-            if let Some(sched_cfg) = sched_events {
-                match crate::telemetry::cpu_profile::SchedProfiler::new(sched_cfg) {
-                    Ok(sched) => shared.push_source(Box::new(sched)),
-                    Err(e) => rate_limited!(Duration::from_secs(60), {
-                        tracing::warn!("failed to start scheduler event profiler: {e}");
-                    }),
-                }
-            }
-        }
-
-        // Channel for TelemetryHandle/Guard → flush thread communication.
-        let (control_tx, control_rx) =
-            crate::primitives::sync::mpsc::sync_channel::<ControlCommand>(1);
-
-        let flush_metrics_sink = worker_metrics_sink
-            .clone()
-            .unwrap_or_else(metrique_writer::sink::DevNullSink::boxed);
-
-        let flush_thread = {
-            let shared = shared.clone();
-            crate::primitives::thread::spawn_named("dial9-flush", move || {
-                #[cfg(target_os = "linux")]
-                // SAFETY: nice() is a simple syscall with no memory safety
-                // implications. Increasing the nice value (lowering priority)
-                // is always permitted for unprivileged processes.
-                unsafe {
-                    let _ = libc::nice(10);
-                }
-
-                #[cfg(feature = "cpu-profiling")]
-                let _ = dial9_perf_self_profile::register_current_thread();
-                run_flush_loop(control_rx, &shared, &flush_metrics_sink, event_writer);
-                #[cfg(feature = "cpu-profiling")]
-                dial9_perf_self_profile::unregister_current_thread();
-            })
-        };
-
-        // Auto-construct worker config when we have a trace path and
-        // at least one processor. When the user supplies no processors
-        // there is nothing for the worker to do, so skip spawning it.
-        let worker_config = trace_path.and_then(|trace_path| {
-            if processors.is_empty() {
-                return None;
-            }
-
-            let poll_interval =
-                worker_poll_interval.unwrap_or(crate::background_task::DEFAULT_POLL_INTERVAL);
-            let metrics_sink =
-                worker_metrics_sink.unwrap_or_else(metrique_writer::sink::DevNullSink::boxed);
-
-            Some(
-                crate::background_task::BackgroundTaskConfig::builder()
-                    .trace_path(trace_path)
-                    .poll_interval(poll_interval)
-                    .processors(processors)
-                    .metrics_sink(metrics_sink)
-                    .build(),
-            )
-        });
-
-        #[allow(unused_mut)]
-        let mut worker = None;
-        if let Some(config) = worker_config {
-            let (shutdown_tx, shutdown_rx) = tokio::sync::oneshot::channel();
-            let wt = crate::primitives::thread::spawn_named("dial9-worker", move || {
-                #[cfg(feature = "cpu-profiling")]
-                let _ = dial9_perf_self_profile::register_current_thread();
-                crate::background_task::run_background_task(config, shutdown_rx);
-                #[cfg(feature = "cpu-profiling")]
-                dial9_perf_self_profile::unregister_current_thread();
-            });
-            worker = Some(WorkerHandle {
-                shutdown: Some(shutdown_tx),
-                thread: Some(wt),
-            });
-        }
-
-        Ok(TelemetryGuard::enabled(
-            TelemetryHandle::enabled(shared, control_tx),
-            Some(flush_thread),
-            worker,
-        ))
-    }
-}
-
-// Custom methods on the generated builder.
-impl<W: TraceWriter, S: telemetry_core_builder::State> TelemetryCoreBuilder<W, S> {
-    /// Configure S3 upload for sealed trace segments.
-    #[cfg(feature = "worker-s3")]
-    pub fn s3_config(mut self, config: crate::background_task::s3::S3Config) -> Self {
-        self.s3_config = Some(config);
-        self
-    }
-
-    /// Provide a pre-built S3 client (for custom credentials or endpoints).
-    #[cfg(feature = "worker-s3")]
-    pub fn s3_client(mut self, client: aws_sdk_s3::Client) -> Self {
-        self.s3_client = Some(client);
-        self
-    }
-
-    /// Set the processor pipeline directly.
-    pub fn processors(
-        mut self,
-        processors: Vec<Box<dyn crate::background_task::SegmentProcessor>>,
-    ) -> Self {
-        self.processors = processors;
-        self
-    }
-
-    /// Set static segment metadata.
-    pub fn segment_metadata(mut self, entries: Vec<(String, String)>) -> Self {
-        self.segment_metadata = entries;
-        self
-    }
-}
-
-/// The flush thread main loop. Extracted so `TelemetryCore::builder` stays readable.
-fn run_flush_loop(
-    control_rx: crate::primitives::sync::mpsc::Receiver<ControlCommand>,
-    shared: &SharedState,
-    flush_metrics_sink: &metrique_writer::BoxEntrySink,
-    mut event_writer: EventWriter,
-) {
-    // Drain the flush thread's own TL buffer every ~1s (200 × 5ms)
-    // rather than every cycle, so queue samples and CPU events
-    // are batched into reasonably-sized segments.
-    let mut cycle_count: u64 = 0;
-    const SELF_DRAIN_INTERVAL: u64 = 200;
-
-    let sample_interval = Duration::from_millis(10);
-    let mut last_sample = time_source().instant();
-    // Snapshot the user-provided segment metadata so we can
-    // merge it with runtime→worker entries on each flush cycle.
-    let static_metadata = event_writer.segment_metadata().to_vec();
-
-    let mut drain_state = DrainState::Idle;
-
-    loop {
-        let mut ack_tx = None;
-        let mut exit = false;
-        // Wait for control commands up to 5ms.
-        match control_rx.recv_timeout(Duration::from_millis(5)) {
-            Ok(ControlCommand::FinalizeAndStop(ack)) => {
-                ack_tx = Some(ack);
-                exit = true;
-            }
-            Err(crate::primitives::sync::mpsc::RecvTimeoutError::Disconnected) => {
-                // All senders dropped — do a best-effort finalize.
-                exit = true;
-            }
-            Err(crate::primitives::sync::mpsc::RecvTimeoutError::Timeout) => {}
-        }
-
-        // When disabled, skip all recording work (queue sampling, metadata
-        // merging, drain coordination, flush). The loop still wakes every
-        // 5ms to check for control commands and the exit signal.
-        if !exit && !shared.is_enabled() {
-            continue;
-        }
-
-        if last_sample.elapsed() >= sample_interval {
-            last_sample = time_source().instant();
-            let contexts = shared.contexts.lock().unwrap().clone();
-            let total_global_queue: usize = contexts.iter().map(|c| c.global_queue_depth()).sum();
-            if !contexts.is_empty() {
-                shared.record_queue_sample(total_global_queue);
-            }
-        }
-
-        // Merge user-provided metadata with runtime→worker mappings
-        // so the next rotated segment is fully self-describing.
-        let contexts = shared.contexts.lock().unwrap().clone();
-        let runtime_entries: Vec<(String, String)> =
-            contexts.iter().filter_map(|c| c.metadata_entry()).collect();
-        if !runtime_entries.is_empty() {
-            let mut merged = static_metadata.clone();
-            merged.extend(runtime_entries);
-            event_writer.update_segment_metadata(merged);
-        }
-
-        cycle_count += 1;
-        let drain_self = exit || cycle_count.is_multiple_of(SELF_DRAIN_INTERVAL);
-        // --- Drain coordination state machine ---
-        //
-        // When the writer reports a drain is due, we can't act immediately
-        // because thread-local buffers may still hold events that belong
-        // in the current segment.  The two-state machine ensures we:
-        //   Idle        → detect should_drain, bump epoch, transition
-        //   EpochBumped → intrusive drain + flush + drained(), back to Idle
-        //
-        // This avoids the bug where re-checking should_drain() every
-        // cycle (it stays true until we actually call drained()) would
-        // forever reschedule the drain and never reach the drained step.
-        let do_drain = match drain_state {
-            DrainState::Idle => {
-                if !exit && event_writer.should_drain() {
-                    shared.bump_drain_epoch();
-                    drain_state = DrainState::EpochBumped;
-                }
-                false
-            }
-            DrainState::EpochBumped => {
-                drain_state = DrainState::Idle;
-                true
-            }
-        };
-
-        // On exit, bump + drain in the same tick since there is no next
-        // tick for the grace period.
-        if exit {
-            shared.bump_drain_epoch();
-        }
-
-        // --- Execute intrusive drain when needed ---
-        if exit || do_drain {
-            let mut tl_drain_timer = Timer::start_now();
-            let stats = shared.drain_all_tl_buffers();
-            tl_drain_timer.stop();
-            let _guard = TlDrainMetrics {
-                operation: Operation::TlDrain,
-                duration: tl_drain_timer,
-                stats,
-                last_drain: exit,
-            }
-            .append_on_drop(flush_metrics_sink.clone());
-        }
-        let mut flush_timer = Timer::start_now();
-        let stats = flush_once(&mut event_writer, shared, drain_self);
-        flush_timer.stop();
-
-        // Notify the writer that TL buffers have been drained and flushed.
-        // The writer may rotate the segment or just advance its drain timer.
-        // Skip on exit — finalize() below will seal the final segment.
-        if do_drain
-            && !exit
-            && let Err(e) = event_writer.drained()
-        {
-            rate_limited!(Duration::from_secs(60), {
-                tracing::warn!("failed to complete post-drain action: {e}");
-            });
-        }
-
-        // Create the metrics guard up front; mutate on the exit path,
-        // then let it drop (which emits the entry).
-        let mut flush_guard =
-            (stats.event_count > 0 || stats.dropped_batches > 0 || exit).then(|| {
-                FlushMetrics {
-                    operation: Operation::Flush,
-                    stats,
-                    flush_duration: flush_timer,
-                    last_flush: exit,
-                    write_metadata_failed: false,
-                    finalize_failed: false,
-                }
-                .append_on_drop(flush_metrics_sink.clone())
-            });
-        if exit {
-            // Write final metadata before sealing so single-segment
-            // traces contain runtime→worker mappings.
-            if let Err(e) = event_writer.write_current_segment_metadata() {
-                rate_limited!(Duration::from_secs(60), {
-                    tracing::warn!("failed to write final segment metadata: {e}");
-                });
-                if let Some(g) = flush_guard.as_mut() {
-                    g.write_metadata_failed = true;
-                }
-            }
-            if let Err(e) = event_writer.finalize() {
-                rate_limited!(Duration::from_secs(60), {
-                    tracing::warn!("failed to finalize trace segment: {e}");
-                });
-                if let Some(g) = flush_guard.as_mut() {
-                    g.finalize_failed = true;
-                }
-            }
-        }
-        drop(flush_guard);
-        if let Some(tx) = ack_tx.take() {
-            let _ = tx.send(());
-        }
-        if exit {
-            return;
-        }
-    }
-}
-
-/// A tokio runtime paired with its (optional) dial9 telemetry guard.
-///
-/// The guard, when present, must outlive the runtime so traces are flushed
-/// on drop — keeping both inside one struct enforces that ordering at the
-/// type level (fields drop top-to-bottom, so `runtime` drops before `guard`).
-///
-/// Construct one of two ways:
-///
-/// - **High-level**: from a [`crate::Dial9Config`] via [`TracedRuntime::new`]
-///   (panicking, used by the `#[dial9_tokio_telemetry::main]` macro) or
-///   [`TracedRuntime::try_new`] (fallible).
-/// - **Low-level**: via [`TracedRuntime::builder`] →
-///   [`TracedRuntimeBuilder::build_and_start`] for direct control over the
-///   raw [`tokio::runtime::Builder`] and [`crate::telemetry::TraceWriter`].
-///   This is the path used by example code, benchmarks, and integration
-///   tests that want to wire a [`crate::telemetry::NullWriter`] or other
-///   custom writer.
-#[derive(Debug)]
-pub struct TracedRuntime {
-    pub(crate) runtime: tokio::runtime::Runtime,
-    pub(crate) guard: TelemetryGuard,
-}
-
-impl TracedRuntime {
-    /// Create a new [`TracedRuntimeBuilder`].
-    pub fn builder() -> TracedRuntimeBuilder<NoTracePath, PipelineUnset> {
-        TracedRuntimeBuilder {
-            enabled: true,
-            task_tracking_enabled: false,
-            task_dump_config: None,
-            trace_path: None,
-            runtime_name: None,
-            #[cfg(feature = "cpu-profiling")]
-            cpu_profiling_config: None,
-            #[cfg(feature = "cpu-profiling")]
-            sched_event_config: None,
-            pipeline: PipelineConfig::Unset,
-            segment_metadata: Vec::new(),
-            worker_poll_interval: None,
-            worker_metrics_sink: None,
-            _marker: std::marker::PhantomData,
-        }
-    }
-
-    /// Build a plain runtime with no telemetry installed.
-    ///
-    /// The returned [`TelemetryGuard`] is in its disabled mode — see
-    /// [`TelemetryGuard::is_enabled`].
-    pub fn build_disabled(
-        mut builder: tokio::runtime::Builder,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        let runtime = builder.build()?;
-        Ok((runtime, TelemetryGuard::disabled()))
-    }
-
-    /// Build the traced runtime. Recording starts disabled.
-    pub fn build(
-        builder: tokio::runtime::Builder,
-        writer: impl TraceWriter + 'static,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        Self::builder().build_with_writer(builder, writer)
-    }
-
-    /// Build the traced runtime and immediately enable recording.
-    pub fn build_and_start(
-        builder: tokio::runtime::Builder,
-        writer: impl TraceWriter + 'static,
-    ) -> std::io::Result<(tokio::runtime::Runtime, TelemetryGuard)> {
-        Self::builder().build_and_start_with_writer(builder, writer)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// High-level construction: TracedRuntime::new / try_new from Dial9Config
-// ---------------------------------------------------------------------------
-//
-// These are the entry points the `#[dial9_tokio_telemetry::main]` macro and
-// hand-written `main` functions reach for. The low-level
-// `TracedRuntime::builder()` / `build_and_start()` above stay available for
-// callers (benches, integration tests, custom-writer setups) that need to
-// drive the [`tokio::runtime::Builder`] directly.
-
-/// Errors produced while constructing a [`TracedRuntime`] from a
-/// [`crate::Dial9Config`].
-///
-/// Writer-transport I/O has already been validated by
-/// [`crate::Dial9ConfigBuilder::build`], so the only remaining failure
-/// modes here come from the tokio runtime builder and the telemetry
-/// background worker startup.
-#[derive(Debug)]
-#[non_exhaustive]
-pub enum TelemetryRuntimeError {
-    /// Failure from [`tokio::runtime::Builder::build`].
-    TokioRuntimeBuilder(std::io::Error),
-    /// Failure from telemetry core setup (traced runtime + background worker).
-    TelemetryCore(std::io::Error),
-}
-
-impl std::fmt::Display for TelemetryRuntimeError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TelemetryRuntimeError::TokioRuntimeBuilder(e) => {
-                write!(f, "tokio runtime builder: {e}")
-            }
-            TelemetryRuntimeError::TelemetryCore(e) => write!(f, "telemetry core: {e}"),
-        }
-    }
-}
-
-impl std::error::Error for TelemetryRuntimeError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        match self {
-            TelemetryRuntimeError::TokioRuntimeBuilder(e)
-            | TelemetryRuntimeError::TelemetryCore(e) => Some(e),
-        }
-    }
-}
-
-/// Drive a [`crate::current_config::Inner`] to a tokio runtime + guard.
-///
-/// `Inner::Enabled` already carries a built [`RotatingWriter`], so this
-/// only needs to materialize the tokio builder and hand both off to
-/// [`TracedRuntimeBuilder::build_and_start`]. `Inner::Disabled`
-/// produces a plain tokio runtime paired with a disabled
-/// [`TelemetryGuard`].
-fn try_assemble_dial9_config(
-    inner: crate::current_config::Inner,
-) -> Result<(tokio::runtime::Runtime, TelemetryGuard), TelemetryRuntimeError> {
-    use crate::current_config::{Inner, materialize_tokio_builder};
-
-    match inner {
-        Inner::Enabled {
-            writer,
-            tokio_configurators,
-            runtime_builder,
-        } => {
-            let tokio_builder = materialize_tokio_builder(&tokio_configurators);
-            let (runtime, guard) = runtime_builder
-                .build_and_start(tokio_builder, writer)
-                .map_err(TelemetryRuntimeError::TelemetryCore)?;
-            Ok((runtime, guard))
-        }
-        Inner::Disabled {
-            tokio_configurators,
-        } => {
-            let runtime = materialize_tokio_builder(&tokio_configurators)
-                .build()
-                .map_err(TelemetryRuntimeError::TokioRuntimeBuilder)?;
-            Ok((runtime, TelemetryGuard::disabled()))
-        }
-    }
-}
-
-impl TracedRuntime {
-    /// Build a [`TracedRuntime`] from a config, panicking with the
-    /// underlying error on failure. Used by the
-    /// `#[dial9_tokio_telemetry::main]` macro.
-    ///
-    /// Reach for this directly when the macro doesn't fit — e.g. when an
-    /// application owns multiple tokio runtimes, when you need to control
-    /// runtime lifetime explicitly, or when you want to drive
-    /// [`TelemetryGuard::graceful_shutdown`] before the runtime drops.
-    ///
-    /// Generic over any input that converts into a [`TracedRuntime`]: in
-    /// practice that means either the fluent
-    /// [`crate::Dial9Config`] (returned by
-    /// [`Dial9Config::builder`](crate::Dial9Config::builder)) or the
-    /// deprecated positional [`crate::config::Dial9Config`]. The generic
-    /// shape is what keeps the macro source-compatible across these
-    /// input types.
-    ///
-    /// # Panics
-    ///
-    /// Panics if the underlying conversion fails — i.e. if the tokio
-    /// runtime cannot be built or the telemetry background worker fails
-    /// to start. When constructing from the fluent
-    /// [`crate::Dial9Config`], writer-transport I/O has already been
-    /// validated by
-    /// [`Dial9ConfigBuilder::build`](crate::Dial9ConfigBuilder::build),
-    /// so the only remaining failure modes are tokio-builder and
-    /// telemetry-core startup I/O.
-    ///
-    /// For fallible construction, use [`try_new`](Self::try_new).
-    ///
-    /// ```no_run
-    /// use dial9_tokio_telemetry::{Dial9Config, TracedRuntime};
-    /// let cfg = Dial9Config::builder()
-    ///     .base_path("trace.bin")
-    ///     .max_file_size(64 * 1024 * 1024)
-    ///     .max_total_size(1024 * 1024 * 1024)
-    ///     .build()?;
-    /// let rt = TracedRuntime::new(cfg);
-    /// rt.block_on(async { /* ... */ });
-    /// # Ok::<(), Box<dyn std::error::Error>>(())
-    /// ```
-    pub fn new<C>(config: C) -> Self
-    where
-        C: TryInto<TracedRuntime>,
-        <C as TryInto<TracedRuntime>>::Error: std::fmt::Display,
-    {
-        config
-            .try_into()
-            .unwrap_or_else(|e| panic!("failed to initialize runtime: {e}"))
-    }
-
-    /// Fallible counterpart to [`new`](Self::new).
-    ///
-    /// Returns the conversion error directly: when constructing from
-    /// [`crate::Dial9Config`] that's a [`TelemetryRuntimeError`]; when
-    /// constructing from the deprecated [`crate::config::Dial9Config`]
-    /// it's a [`std::io::Error`]. Use this when you want to handle
-    /// runtime construction failure rather than panic.
-    ///
-    /// ```no_run
-    /// use dial9_tokio_telemetry::{Dial9Config, TracedRuntime};
-    /// let cfg = Dial9Config::builder()
-    ///     .base_path("trace.bin")
-    ///     .max_file_size(64 * 1024 * 1024)
-    ///     .max_total_size(1024 * 1024 * 1024)
-    ///     .build()?;
-    /// let rt = TracedRuntime::try_new(cfg)?;
-    /// rt.block_on(async { /* ... */ });
-    /// # Ok::<(), Box<dyn std::error::Error>>(())
-    /// ```
-    pub fn try_new<C>(config: C) -> Result<Self, <C as TryInto<TracedRuntime>>::Error>
-    where
-        C: TryInto<TracedRuntime>,
-    {
-        config.try_into()
-    }
-
-    /// Borrow the underlying tokio runtime.
-    pub fn runtime(&self) -> &tokio::runtime::Runtime {
-        &self.runtime
-    }
-
-    /// Borrow the telemetry guard.
-    ///
-    /// The guard is always present, regardless of whether telemetry was
-    /// installed. Use [`TelemetryGuard::is_enabled`] to distinguish a
-    /// live telemetry session from an inert (disabled) guard.
-    pub fn guard(&self) -> &TelemetryGuard {
-        &self.guard
-    }
-
-    /// Run `fut` to completion on the runtime.
-    ///
-    /// The future is always spawned through the guard's
-    /// [`TelemetryHandle`]. On an enabled guard this records poll and
-    /// wake events; on a disabled guard the handle's `spawn` falls
-    /// through to plain [`tokio::spawn`].
-    pub fn block_on<F>(&self, fut: F) -> F::Output
-    where
-        F: std::future::Future + Send + 'static,
-        F::Output: Send + 'static,
-    {
-        let handle = self.guard.handle();
-        self.runtime.block_on(async move {
-            match handle.spawn(fut).await {
-                Ok(output) => output,
-                Err(err) if err.is_panic() => std::panic::resume_unwind(err.into_panic()),
-                Err(_) => unreachable!("task cannot be cancelled inside block_on"),
-            }
-        })
-    }
-}
-
-impl TryFrom<crate::Dial9Config> for TracedRuntime {
-    type Error = TelemetryRuntimeError;
-
-    fn try_from(config: crate::Dial9Config) -> Result<Self, Self::Error> {
-        let (runtime, guard) = try_assemble_dial9_config(config.0)?;
-        Ok(Self { runtime, guard })
-    }
-}
-
-/// Bridge for the deprecated positional config API at
-/// [`crate::config::Dial9Config`] so that it remains compatible with
-/// [`TracedRuntime::new`] (and therefore the
-/// `#[dial9_tokio_telemetry::main]` macro).
-impl TryFrom<crate::config::Dial9Config> for TracedRuntime {
-    type Error = std::io::Error;
-
-    fn try_from(config: crate::config::Dial9Config) -> Result<Self, Self::Error> {
-        let (runtime, guard) = config.build()?;
-        Ok(Self {
-            runtime,
-            guard: guard.unwrap_or_else(TelemetryGuard::disabled),
-        })
-    }
-}
-
 #[cfg(all(test, not(shuttle)))]
 mod tests {
     use super::*;
     use crate::telemetry::NullWriter;
+    use crate::telemetry::buffer;
     use crate::telemetry::collector::CentralCollector;
+    use crate::telemetry::writer::RotatingWriter;
     use std::panic::Location;
     use std::sync::Arc;
     use std::sync::atomic::AtomicU64;
